% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggplot.R
\name{annotate_label}
\alias{annotate_label}
\title{Add a label in the corner of an existing ggplot}
\usage{
annotate_label(
  plt,
  label,
  location = c(0.02, 0.98),
  justification = c(0, 1),
  label_spacing = 0.1,
  size = 3,
  color = "black",
  parse = FALSE
)
}
\arguments{
\item{plt}{ggplot object}

\item{label}{vector with character strings to add}

\item{location}{two-part vector indicating the relative location in the
plot for placing labels. Follows same logic as legend placement in ggplot.}

\item{justification}{two-part vector indicating the horizontal and
vertical justification of the labels. Follwos the same logic as legend
justification in ggplot}

\item{label_spacing}{if `label` consists of multiple elements, the relative
`y`-spacing between each element}

\item{size}{plot size of text}

\item{color}{plot color of text}

\item{parse}{if `TRUE`, labels are parsed when plotted}
}
\value{
ggplot object with added annotation
}
\description{
Add a text label an existing ggplot object. For example. use to annotate
a subplot indicator to an exsiting plot
}
