% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomech.R
\name{biomech_fit}
\alias{biomech_fit}
\title{Wrapper function for fitting biomechanical parameters}
\usage{
biomech_fit(
  dr,
  epsru,
  tru,
  dr0 = 1,
  weights_power = dr^2,
  weights_copula = dr^2
)
}
\arguments{
\item{dr, epsru, tru}{vectors with measured diameters, tensile strain to peak
and tensile strength}

\item{dr0}{root reference diameter}

\item{weights_power}{array with weights to use for power-law fitting}

\item{weights_copula}{array with weights to use for copula fitting}
}
\value{
a dataframe with key parameters for tensile strain to peak fitting
  (power law parameters `epsru0` and `betaeps`, Weibull shape parameter
  `kappaeps`), tensile strength fitting (power law parameters `tru0` and
  `betat`, Weibull shape parameter `kappat`) and copula fitting (
  correlation coefficient `rho`)
}
\description{
Generate fits for root tensile strain to failure and tensile strength
(+ correlation between normalised strain and strength). Uses power-law
fitting function `power_weibull_fit()` for power-law fitting and the
function `copula_gaussian_fit()` to fit the correlation coefficient
for the copula describnig normalised strain and strength relationships.
}
\examples{
# generate data
n <- 101
dr <- seq(1, 7, l = n)
df <- biomech_random(dr, 0.2, -0.2, 4, 10, -0.5, 6, 0.5)

# fit
ft <- biomech_fit(df$dr, df$epsru, df$tru)

# power-law predictions
dr_pred <- seq(min(dr), max(dr), l = 251)
epsru_pred <- ft$epsru0*dr_pred^ft$betaeps
tru_pred <- ft$tru0*dr_pred^ft$betat

# prediction interval
df_ell <- bivariate_normal_predictioninterval(rho = ft$rho)
df_pred <- bivariate_normal_toweibull(df_ell$x, df_ell$y, ft$kappaeps, ft$kappat)

# plot - tensile strain
plot(dr, df$epsru)
lines(dr_pred, epsru_pred)

# plot - tensile strength
plot(dr, df$tru)
lines(dr_pred, tru_pred)

# plot - relationship epsru/tru
plot(
  df$epsru/(ft$epsru0*dr^ft$betaeps),
  df$tru/(ft$tru0*dr^ft$betat)
)
lines(df_pred$x, df_pred$y)

}
