% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomech.R
\name{biomech_random}
\alias{biomech_random}
\title{Generate random root properties given biomechanical fit}
\usage{
biomech_random(
  dr,
  epsru0,
  betaeps,
  kappaeps,
  tru0,
  betat,
  kappat,
  rho,
  dr0 = 1
)
}
\arguments{
\item{dr}{array with known root diameters}

\item{epsru0}{diameter-tensile strain power-law multiplier}

\item{betaeps}{diameter-tensile strain power-law power coefficient}

\item{kappaeps}{diameter-tensile strain Weibull shape parameter}

\item{tru0}{diameter-tensile strength power-law multiplier}

\item{betat}{diameter-tensile strength power-law power coefficient}

\item{kappat}{diameter-tensile strength Weibull shape parameter}

\item{rho}{copula correlation coefficient}

\item{dr0}{root reference diameter}
}
\value{
dataframe with root diameters (`dr`), tensile strain to peak
(`epsru`) and tensile strength (`tru`)
}
\description{
Generate random biomechanical properties (strain to failure, strength) for
known root diameters, using biomechanical fitting parameters
}
\examples{
# generate data
dr <- seq(2, 9, l = 100)
df <- biomech_random(dr, 0.2, -0.3, 4, 10, -0.2, 6, 0.4)

# plot
plot(df$dr, df$epsru)
plot(df$dr, df$tru)

}
