% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_normal.R
\name{bivariate_normal_axes}
\alias{bivariate_normal_axes}
\title{Major and minor axes for bivariate normal confidence ellipsoid}
\usage{
bivariate_normal_axes(
  mu = c(0, 0),
  sd = c(1, 1),
  rho = 0,
  confidence = 0.95,
  n = 100
)
}
\arguments{
\item{mu}{vector with means for x and y}

\item{sd}{vector with standard deviations for x and y}

\item{rho}{correlation coefficient for the bivariate normal distribution}

\item{confidence}{confidence value of the ellipsoid}

\item{n}{number of points to use for each axis}
}
\value{
dataframe with axis type (fields `axis`, can be `major` or
  `minor`), and the coordiante fields `x` and `y`
}
\description{
Generate coordinates to plot the minor and major axes of the
bivariate normal confidence ellipsoid
}
\examples{
bivariate_normal_axes(rho = 0.5, n = 10)

}
