% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_normal.R
\name{bivariate_normal_density}
\alias{bivariate_normal_density}
\title{Probability density for a bivariate normal distribution}
\usage{
bivariate_normal_density(x, mu = c(0, 0), sd = c(1, 1), rho = 0, log = FALSE)
}
\arguments{
\item{x}{n*2 matrix with observations}

\item{mu}{vector with means for x and y}

\item{sd}{vector with standard deviations for x and y}

\item{rho}{correlation coefficient for the bivariate normal distribution}

\item{log}{if `log = TRUE`, return the log-transformed probability density
rather than the probability density}
}
\value{
vector with probability density for each observation
}
\description{
Probability density for a bivariate normal distribution
}
\examples{
# input
mu <- c(5, 2)
sd <- c(2, 1)
rho <- -0.6

# generate grid
x <- mu[1] + 3*sd[1]*seq(-1, 1, l = 101)
y <- mu[2] + 3*sd[2]*seq(-1, 1, l = 101)
xy <- expand.grid(x = x, y = y, KEEP.OUT.ATTRS = FALSE)

# calculate probability density
d <- matrix(
  bivariate_normal_density(xy, mu = mu, sd = sd, rho = rho),
  nrow = length(x),
  ncol = length(y)
)

# plot contour plot
contour(x, y, d)

}
