% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_normal.R
\name{bivariate_normal_predictioninterval}
\alias{bivariate_normal_predictioninterval}
\title{Generate prediction interval for bivariate normal distribution}
\usage{
bivariate_normal_predictioninterval(
  mu = c(0, 0),
  sd = c(1, 1),
  rho = 0,
  confidence = 0.95,
  n = 360
)
}
\arguments{
\item{mu}{vector with means for x and y}

\item{sd}{vector with standard deviations for x and y}

\item{rho}{correlation coefficient for the bivariate normal distribution}

\item{confidence}{value of confidence required, centred around the mean (
i.e. setting `confidence = 0.95` gives the ellipsoid that contains a random
point with 95\% confidence, with tails of 2.5\%)}

\item{n}{number of points on ellipsoid}
}
\value{
dataframe with coordinates `x` and `y` for each confidence level
  in `confidence`
}
\description{
Generate the coordinates of the prediction interval ellipsoid for a
bivariate normal distribution.
}
\examples{
# parameters
mu <- c(2, 5)
sd <- c(1, 0.2)
rho <- 0.9
confidence <- 0.95

# generate some random data
df <- bivariate_normal_random(100, mu = mu, sd = sd, rho = rho)
# generate confidence interval
dc <- bivariate_normal_predictioninterval(mu = mu, sd = sd, rho = rho)

# plot
plot(df$x, df$y)
lines(dc$x, dc$y)

}
