% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_normal.R
\name{bivariate_normal_random}
\alias{bivariate_normal_random}
\title{Generate random bivariate normal data}
\usage{
bivariate_normal_random(n, mu = c(0, 0), sd = c(1, 1), rho = 0)
}
\arguments{
\item{n}{number of samples to draw}

\item{mu}{vector with means for x and y}

\item{sd}{vector with standard deviations for x and y}

\item{rho}{correlation coefficient for the bivariate normal distribution}
}
\value{
dataframe with `x` and `y` fields for randomly generated data
}
\description{
Generate random bivariate normally distributed data, with known means,
standard deviations and a correlation coefficient.
}
\examples{
# example1: mean = 0, sd = 1
x <- bivariate_normal_random(100)
plot(x[, 1], x[, 2])

# correlated data, with offset
df <- bivariate_normal_random(100, mu = c(3, 6), sd = c(1, 2), rho = -0.5)
plot(df$x, df$y)

}
