% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_normal.R
\name{bivariate_normal_toweibull}
\alias{bivariate_normal_toweibull}
\title{Convert bivariate normal values to weibull}
\usage{
bivariate_normal_toweibull(
  x,
  y,
  shape_x,
  shape_y,
  mean = c(0, 0),
  sd = c(1, 1),
  scale_x = 1/gamma(1 + 1/shape_x),
  scale_y = 1/gamma(1 + 1/shape_y)
)
}
\arguments{
\item{x, y}{positions in bivariate normal coordinates}

\item{shape_x, shape_y}{Weibull shape parameters for x and y}

\item{sd}{vector with standard deviations for x and y}

\item{scale_x, scale_y}{Weibull scale parameters for x and y}
}
\value{
dataframe with Weibull positions `x` and `y`
}
\description{
Convert bivariate normally distributed values to weibull-distributed
values. This is done by setting the cumulative densities of both
distributions the same
}
\examples{
df_norm <- bivariate_normal_predictioninterval(rho = 0.5)
plot(df_norm$x, df_norm$y, "l")

df_weibull <- bivariate_normal_toweibull(
  df_norm$x, df_norm$y, 4, 6, scale_y = 4)
plot(df_weibull$x, df_weibull$y, "l")
}
