% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_normal.R
\name{bivariate_normal_variancematrix}
\alias{bivariate_normal_variancematrix}
\title{Variance-covariance matrix for bivariate normal distribution}
\usage{
bivariate_normal_variancematrix(sd = c(1, 1), rho = 0)
}
\arguments{
\item{sd}{vector with standard deviations}

\item{rho}{correlation coefficient}
}
\value{
2*2 matrix
}
\description{
Generate variance-covariance matrix for a bivariate normal distribution,
based on known standard deviations and correlation coefficient.
}
\examples{
bivariate_normal_variancematrix(sd = c(2, 3), rho = 0.4)

}
