% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{breakpoint_options}
\alias{breakpoint_options}
\title{Select breakpoints for initial root count guess guess}
\usage{
breakpoint_options(n, xmin, xmax, n0 = 4, fixed = rep(NA, n))
}
\arguments{
\item{n}{number of breakpoints to select}

\item{xmin, xmax}{min and max value of domain}

\item{n0}{total number of extra optional locations}

\item{fixed}{array with values of fixed breakpoints. Should contain `NA` for
each breakpoint that is unknown}
}
\value{
matrix with breakpoint options
}
\description{
Generate a series of optional x-breakpoints on a known interval `xmin` to
`xmax`. The domain is split into equal intervals, where `n0` indicates the
extra number of optional points on top of `n`. All combinations of `n`
breakpoints from the optional set with length `n + n0` are returned as a
matrix

Specific breakpoints may be fixed using the `fixed` argument.
}
\examples{
# input
n <- 3
xmin <- 0
xmax <- 10

# no fixes
breakpoint_options(n, xmin, xmax, n0 = 2)
# fix in middle
breakpoint_options(n, xmin, xmax, fixed = c(NA, 6, NA), n0 = 4)
# fix at end
breakpoint_options(n, xmin, xmax, fixed = c(NA, NA, 6), n0 = 4)

}
