% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_perfectfit}
\alias{check_perfectfit}
\title{Check if all points are perfectly on power-law line}
\usage{
check_perfectfit(x, y, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x, y}{vectors of diameter and strength data}

\item{tol}{tolerance}
}
\value{
list with fields:
\itemize{
\item \code{colinear}: boolean that is true when all points are colinear
\item \code{multiplier}: multiplier of power law fit
\item \code{exponent}: multiplier of power law fit
}
}
\description{
Checked by comparing co-linearity in log-log space
}
\examples{
x <- seq(3, 6, l = 5)
y <- 20*x^-0.5
check_perfectfit(x, y)

y[2] <- y[2] + 0.1
check_perfectfit(x, y)

}
