% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confidence_ellipse}
\alias{confidence_ellipse}
\title{Generate confidence ellipse for 2-dimensional data}
\usage{
confidence_ellipse(x, y, level = 0.95, n = 181)
}
\arguments{
\item{x, y}{vectors with data}

\item{level}{confidence level}

\item{n}{number of points to generate on ellipse circumference}
}
\value{
list with fields:
\itemize{
\item \code{ellipse}: dataframe with \code{x} and \code{y} coordinates of the
confidence ellipse
\item \code{major} dataframe with \code{x} and \code{y} coordinates of the major
axis within the ellipse
\item \code{minor} dataframe with \code{x} and \code{y} coordinates of the minor
axis within the ellipse
\item \code{mean}: dataframe with \code{x} and \code{y} coordinates of the mean
}
}
\description{
Generate coordinates for a x\% confidence ellipse.
}
\examples{
x <- rnorm(100, 2, 1)
y <- rnorm(100, 4, 0.5) + 0.5*x
df <- confidence_ellipse(x, y)
plot(x, y)
lines(df$ellipse$x, df$ellipse$y, col = "red")
lines(df$major$x, df$major$y, col = "red")
lines(df$minor$x, df$minor$y, col = "red")
lines(df$mean$x, df$mean$y, type = "p", col = "red")

}
