% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_gaussian.R
\name{copula_gaussian_fit}
\alias{copula_gaussian_fit}
\title{Fit gaussian copula using loglikelihood fitting}
\usage{
copula_gaussian_fit(px, py, weights = rep(1, length(px)), eps = 1e-12)
}
\arguments{
\item{px, py}{vectors with cumulative probability densities}

\item{weights}{vector with weights for each observation}

\item{eps}{small value, so that the correlation coefficient is set on
determined on the interval -(1 - eps) <= rho <= (1 - eps). rho >= 1
and rho <= -1 will lead to numerical problems.}
}
\value{
dataframe with weighted loglikelihood of the best fit (field
  `loglikelihood`) and the optimal value of the correlation coefficient
  (field `rho`)
}
\description{
Find the correlation coefficient for the best gaussian copula
fit through given two-dimensional cumulative density data.
}
\examples{
# generate data
xy <- bivariate_normal_random(1000, rho = -0.6)
px <- pnorm(xy[, 1])
py <- pnorm(xy[, 2])

# fit correlation coefficient
ft <- copula_gaussian_fit(px, py)

# confidence interval
dc <- bivariate_normal_predictioninterval(rho = ft$rho)

# plot observations + prediction interval
plot(xy$x, xy$y)
lines(dc$x, dc$y)

}
