% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_gaussian.R
\name{copula_gaussian_initialguess}
\alias{copula_gaussian_initialguess}
\title{Initial guess for gaussian copula fitting}
\usage{
copula_gaussian_initialguess(px, py)
}
\arguments{
\item{px, py}{vectors with cumulative probability densities}
}
\value{
initial guess for correlation coefficient `rho`
}
\description{
Make an initial estimate for correlation coefficient rho, which is
used in the gaussian copula fitting function `copula_gaussian_fit()`
}
