% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_gaussian.R
\name{copula_gaussian_likelihood}
\alias{copula_gaussian_likelihood}
\title{Objective loglikelihood function for gaussian copula fitting}
\usage{
copula_gaussian_likelihood(rho, px, py, weights = rep(1, length(px)))
}
\arguments{
\item{rho}{correlation coefficient to fit}

\item{px, py}{vectors with cumulative probability densities}

\item{weights}{vector with weights for each observation}
}
\value{
weighted loglikeihood value
}
\description{
Returns scaled version of the loglikelihood of the gaussian copula fit.
There is a single fitting parameter: correlation coefficient `rho`.

The loglikelihood value is given by:

  L = sum(weights*(dx + dy + cm))

where dx and dy are the probability densities of the marginal distributions,
and cm the copula multiplier. Since only cm is dependent on the value
of rho, we can instead maximise instead:

  L = sum(weights*cm)
}
