% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_gaussian.R
\name{copula_gaussian_likelihood_jacobian}
\alias{copula_gaussian_likelihood_jacobian}
\title{Jacobian of the function `copula_gaussian_likelihood()`}
\usage{
copula_gaussian_likelihood_jacobian(rho, px, py, weights = rep(1, length(px)))
}
\arguments{
\item{rho}{correlation coefficient to fit}

\item{px, py}{vectors with cumulative probability densities}

\item{weights}{vector with weights for each observation}
}
\value{
derivative of `copula_gaussian_likelihood()` with respect to `rho`
}
\description{
Returns the derivative of the (scalar) function
`copula_gaussian_likelihood()` with respect to input parameter `rho`
}
\examples{
# Test: compare analytical with numerical jacobian
rho <- 0.6
px <- seq(0.4, 0.5, l = 10)
py <- seq(0.2, 0.4, l = 10)
weights <- seq(0.8, 1.3, l = 10)

eps <- 1e-6
L1 <- copula_gaussian_likelihood(rho, px, py)
L2 <- copula_gaussian_likelihood(rho + eps, px, py)
J2 <- (L2 - L1)/eps
J <- copula_gaussian_likelihood_jacobian(rho, px, py)

c(J, J2)

}
