% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_gaussian.R
\name{copula_gaussian_multiplier}
\alias{copula_gaussian_multiplier}
\title{Gaussian copula multiplier}
\usage{
copula_gaussian_multiplier(px, py, rho = 0, log = FALSE)
}
\arguments{
\item{px, py}{vectors with cumulative probability densities}

\item{rho}{correlation coefficient}

\item{log}{if `log = TRUE`, return the log-transformed copula multipliers
instead}
}
\value{
vector with gaussian copula multipliers for each pair of `px` and
`py`.
}
\description{
Return gaussian copula multipliers based on known 2-d cumulative probability
density values and known correlation coefficient.

The function is vectorised.
}
\examples{
# input
rho <- 0.8
mu <- c(2, 3)
sd <- c(1, 1)

# generate some data
x <- mu[1] + 3*sd[1]*seq(-1, 1, l = 101)
y <- mu[2] + 3*sd[2]*seq(-1, 1, l = 101)

# grid
df <- expand.grid(x = x, y = y, KEEP.OUT.ATTRS = FALSE)
# probability density, and cumulative probability density
df$dx <- dnorm(xy$x, mean = mu[1], sd = sd[1])
df$dy <- dnorm(xy$y, mean = mu[2], sd = sd[2])
df$px <- pnorm(xy$x, mean = mu[1], sd = sd[1])
df$py <- pnorm(xy$y, mean = mu[2], sd = sd[2])

# copula multiplier
df$cm <- copula_gaussian_multiplier(df$px, df$py, rho = rho)

# matrix with probability
z <- matrix(with(df, cm*dx*dy), nrow = length(x), ncol = length(y))

# plot
contour(x, y, z)

}
