% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_gaussian.R
\name{copula_gaussian_multiplier_jacobian}
\alias{copula_gaussian_multiplier_jacobian}
\title{Derivative of `copula_gaussian_multiplier()`}
\usage{
copula_gaussian_multiplier_jacobian(px, py, rho = 0, log = FALSE)
}
\arguments{
\item{px, py}{vectors with cumulative probability densities}

\item{rho}{correlation coefficient}

\item{log}{if `log = TRUE`, return the log-transformed copula multipliers
instead}
}
\value{
list with derivatives, with fieldnames `px`, `py` and `rho`.
}
\description{
Calculate derivative of copula multiplier calculated using the function
`copula_gaussian_multiplier()` with respect to its input parameters `px`,
`py` and `rho`.

The function is vectorised.
}
\examples{
px <- 0.1
py <- 0.2
rho <- -0.8
log <- FALSE

cm0 <- copula_gaussian_multiplier(px, py, rho, log = log)
J <- copula_gaussian_multiplier_jacobian(px, py, rho, log = log)

eps <- 1e-6
J2 <- (c(
  copula_gaussian_multiplier(px + eps, py, rho, log = log),
  copula_gaussian_multiplier(px, py + eps, rho, log = log),
  copula_gaussian_multiplier(px, py, rho + eps, log = log)
) - cm0)/eps

J
J2

}
