% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggplot.R
\name{create_annotations}
\alias{create_annotations}
\title{Create annotation labels for plotting}
\usage{
create_annotations(
  values,
  prefix = NULL,
  suffix = NULL,
  parse = TRUE,
  sep = "",
  digits = 3,
  x = 0.98,
  y = 0.98 - seq(0, length(values) - 1) * 0.1,
  hjust = 1,
  vjust = 1
)
}
\arguments{
\item{values}{array with values to be rounded}

\item{prefix, suffix}{arrays with character prefix and suffixes}

\item{parse}{if `TRUE`, the combined character string is assumed as an
parseable expression}

\item{sep}{seperator to put between prefix, value and suffixes}

\item{digits}{number of significant digits to use when rounding values}

\item{x, y}{relative location of label in the plot (0 <= x,y <= 1)}

\item{hjust, vjust}{horizontal and vertical justification of labels in the
plot}
}
\value{
dataframe with fields for labels (`label`), relative position in
  plot (`x`, `y`), text justification (`hjust`, `vjust`) and whether the
  labels should be parsed or not (`parse`)
}
\description{
Creates character strings for annotating a plot. Combines a prefix, a
numeric value (will be converted to a string using the `numeric2character()`
function) and a character suffix.
}
\examples{
create_annotations(
  c(0.23, -0.1),
  prefix = c("a=", ""),
  suffix = c(" mm", " N"),
  parse = FALSE,
  digits = 4
)

}
