% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{dgumbel}
\alias{dgumbel}
\title{Gumbel probability distribution density}
\usage{
dgumbel(x, location = 0, scale = 1, weights = rep(1, length(x)), log = FALSE)
}
\arguments{
\item{x}{vector with x-values}

\item{location}{Gumbel location parameter}

\item{scale}{Gumbel scale parameter}

\item{weights}{vector with weights for each x-value}

\item{log}{if `TRUE`, return log-transformed probability densities}
}
\value{
vector with probability density for each value in x
}
\description{
Calculate Gumbel probability density
}
\examples{
dgumbel(seq(1, 3, l = 5), location = 2, scale = 2)

}
