% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{erf}
\alias{erf}
\title{Error function}
\usage{
erf(x)
}
\arguments{
\item{x}{input values for the error function}
}
\value{
erf(x)
}
\description{
Calculate the value of the error function
}
\examples{
erf(0.5)
}
