% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{erfi}
\alias{erfi}
\title{Inverse error function}
\usage{
erfi(x)
}
\arguments{
\item{x}{input values for the inverse error function}
}
\value{
erf^{-1}(x)
}
\description{
Calculate the value of the inverse error function
}
\examples{
erfi(0.5)

}
