% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{expand_grid_df}
\alias{expand_grid_df}
\title{expand.grid function for dataframes}
\usage{
expand_grid_df(...)
}
\arguments{
\item{...}{input dataframes}
}
\value{
dataframe with all combinations of rows
}
\description{
Generate all combinations of rows in a series of data.frames. This is the
dataframe equivalent of the standard R function `expand.grid()`, which only
works on vectors.

The code is taken from Stack Overflow:
https://stackoverflow.com/questions/11693599/alternative-to-expand-grid-for-data-frames

Iterations loop fastest through dataframes specified first.
}
\examples{
df1 <- data.frame(a = c(1, 2), b = c(3, 4))
df2 <- data.frame(d = c(1, 2, 3), e = c(4, 5, 6), f = c(7, 8, 9))
expand_grid_df(df1, df2)

}
