% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{gumbel_fit}
\alias{gumbel_fit}
\title{Gumbel likelihood fitting}
\usage{
gumbel_fit(
  x,
  weights = rep(1, length(x)),
  method = "bisection",
  range = c(-1, 1),
  start = NULL,
  scale_min = 0.1
)
}
\arguments{
\item{x}{array with observations}

\item{weights}{weighting factor for each observation in \code{x}}

\item{method}{choose \code{newton} for gradient descent solving, using the
\code{rootSolve::multiroot()} function, or \code{bisection} for bisection root
solving algorithm using the \code{stats::uniroot()} function.}

\item{range}{two-value array to add to best guess, to define initial
interval for bisection algorithm}

\item{start}{initial guess for weibull shape parameter. If not defined, an
estimate is made based on fitting of the linearised cumulative density
distribution}

\item{scale_min}{minimum value of scale parameter. Set to a small value,
since by definition scale > 0}
}
\value{
a list containing the fields
\itemize{
\item \code{loglikelihood}: the log-likelihood score of the fit
\item \code{location}: fitted gumbel location parameter
\item \code{scale}: fitted gumbel scale parameter
}
}
\description{
Determine best fitting Gumbel distribution using weighted loglikelihood
fitting.
}
\examples{
mu <- 20
theta <- 3
x <- rgumbel(20, mu, theta)
gumbel_fit(x)

}
