% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{gumbel_loglikelihood}
\alias{gumbel_loglikelihood}
\title{Calculate Gumbel loglikelihood}
\usage{
gumbel_loglikelihood(par, x, weights = rep(1, length(x)), deriv = 0)
}
\arguments{
\item{par}{vector with fitting parameters (location and scale parameter)}

\item{x}{array with observations}

\item{weights}{weighting factor for each observation in \code{x}}

\item{deriv}{order of partial derivative requested}
}
\value{
loglikelihood, or its partial derivatives to order `deriv`
}
\description{
Calculate the weighted loglikelihood for a gumbel fit.
Can also be used to calculate the first or second partial
derivative with respect to fitting parameters.
}
\keyword{internal}
