% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{gumbel_root}
\alias{gumbel_root}
\title{Root to solve for Gumbel fitting}
\usage{
gumbel_root(theta, x, weights = rep(1, length(x)))
}
\arguments{
\item{theta}{shape parameter}

\item{x}{array with observations}

\item{weights}{weighting factor for each observation in \code{x}}
}
\value{
root
}
\description{
Root equation to solve for Gumbel fitting
}
\keyword{internal}
