% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numeric2character}
\alias{numeric2character}
\title{Round a double to a number of significant digits and convert to character}
\usage{
numeric2character(x, digits = 3, round = FALSE)
}
\arguments{
\item{x}{numeric value to convert to character string}

\item{digits}{number of significant digits}

\item{round}{if `TRUE`, round `x` to `digits` number of decimals. If
`FALSE`, round to `digits` number of significant digits}
}
\value{
character string
}
\description{
Round a number to a number of significant digits and convert to character
string. Always returns `digits` number of significant digits, also when
there are trailing zeros. if `round = TRUE`, the number is rounded to
`digits` number of decimals instead.
}
\examples{
numeric2character(pi, digits = 4)
numeric2character(3, digits = 2, round = TRUE)
numeric2character(sqrt(5)/100000, digits = 3)

}
