% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal_fit.R
\name{orthagonal_fit}
\alias{orthagonal_fit}
\title{Orthagonal linear fitting}
\usage{
orthagonal_fit(x, y)
}
\arguments{
\item{x, y}{vectors with x and y data}
}
\value{
vector with angle (in radians) and offset, describing the fitting
  line. The angle is defined as positive rotating from the x-axis to the
  y-axis, and the offset is defined as the distance between the origin
  and the fitting line
}
\description{
Conduct an orthagonal linear fit, minimising the sum of squared distances
between datapoints and the linear fit
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(2, 3, 3, 5, 5)
ft <- orthagonal_fit(x, y)
xp <- seq(min(x), max(x), l = 11)
yp <- ft[2]/cos(ft[1]) + xp*tan(ft[1])
plot(x, y, xlim = c(0, 6), ylim = c(0, 6))
lines(xp, yp)

}
