% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggplot.R
\name{plot_settings}
\alias{plot_settings}
\title{List with general plot settings}
\usage{
plot_settings(
  color_meas = "#377EB8",
  fill_meas = color_meas,
  size_meas = 1,
  linewidth_meas = 0.5,
  linetype_meas = 1,
  shape_meas = 20,
  alpha_meas = 0.25,
  color_fit = "#E41A1C",
  fill_fit = color_fit,
  linewidth_fit = 0.5,
  linetype_fit = 1,
  shape_fit = 20,
  alpha_fit = 0.25,
  color_ann = color_fit,
  size_ann = 3
)
}
\arguments{
\item{color_meas}{color for measurements}

\item{fill_meas}{fill color for measurements polygons}

\item{size_meas}{point size for measurements}

\item{linewidth_meas}{linewidth for measurements lines}

\item{linetype_meas}{linetype for measurements lines}

\item{shape_meas}{shape for measurement points}

\item{alpha_meas}{transparancy values for measurements polygons}

\item{color_fit}{color for fitting lines}

\item{fill_fit}{fill color for fitting polygons}

\item{linewidth_fit}{linewidth for fitting lines}

\item{linetype_fit}{linetype for fitting lines}

\item{shape_fit}{shape for fitting points}

\item{alpha_fit}{transparancy values for fitting polygons}

\item{color_ann}{color for text annotations}

\item{size_ann}{font size for text annotations}
}
\value{
list with fields (input parameter names) and values
}
\description{
Generate a list with 'standard' plot settings.
}
\examples{
plot_settings()

}
