% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_integrate.R
\name{power_integrate}
\alias{power_integrate}
\title{Integrate a power-law curve}
\usage{
power_integrate(power, lower, upper, multiplier = 1)
}
\arguments{
\item{power}{vector with powers}

\item{lower, upper}{vectors with lower and upper integration limits}

\item{multiplier}{(optional) vector with multipliers}
}
\value{
vector with integrands
}
\description{
Integrate a power law curve in the form:

  y(x) = a*x^b

between x=x1 and x=x2.

Function is vectorised.
}
\examples{
power_integrate(2.1, 2, 5, multiplier = 3)
stats::integrate(function(x) 3*x^2.1, 2, 5)

}
