% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{power_integrate_jacobian}
\alias{power_integrate_jacobian}
\title{Derivative of function `power_integrate()`}
\usage{
power_integrate_jacobian(b, x1, x2)
}
\arguments{
\item{b}{vector with powers}

\item{x1, x2}{vectors with lower and upper integration limits}
}
\value{
a list containing the derivatives for each input parameter. Has
  fields `b`, `x1` and `x2`
}
\description{
Generates the derivative of the results of the function `power_integrate()`
with respect to its input parameters `b`, `x1` and `x2`.

Function is vectorised.
}
\examples{
# Test jacobian by comparing to numberical solution
b <- 1.2
x1 <- 2.1
x2 <- 3.3

I <- power_integrate(b, x1, x2)
eps <- 1e-6
J <- power_integrate_jacobian(b, x1, x2)

Jb <- (power_integrate(b + eps, x1, x2) - I)/eps
Jx1 <- (power_integrate(b, x1 + eps, x2) - I)/eps
Jx2 <- (power_integrate(b, x1, x2 + eps) - I)/eps

J
c(Jb, Jx1, Jx2)

}
