% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_loglikelihood}
\alias{power_loglikelihood}
\title{Loglikelihood function for power law probability density}
\usage{
power_loglikelihood(
  b,
  x,
  xmin = min(x),
  xmax = max(x),
  weights = rep(1, length(x)),
  eps = .Machine$double.eps^0.5
)
}
\arguments{
\item{b}{power coefficient}

\item{x}{vector with observations (must all be finite and >0)}

\item{xmin, xmax}{min and max limits of the density function}

\item{weights}{weighting for each observation in \code{x}}

\item{eps}{small value, to catch case where b == -1}
}
\value{
loglikelihood score
}
\description{
Returns the loglikelihood score for a power law probability density
function, given the power coefficient `b`.
}
\examples{
b <- 2
x <- seq(1, 7, l = 101)^(1/b)

eps <- 1e-6
f0 <- power_loglikelihood(b, x)
f1 <- power_loglikelihood(b + eps, x)
(f1 - f0)/eps
power_root(b, x)

}
