% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_nls.R
\name{power_nls_fit}
\alias{power_nls_fit}
\title{Fit power-law function using least-squares regression}
\usage{
power_nls_fit(x, y, weights = rep(1, length(x)), guess = NULL)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a unitless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{weighting for each measurement. Default = 1, but a strong
case can be made for weighting with `weights = x^2` because of the
large effect of thick roots on root-reinforcement}

\item{guess}{vector with initial guess for the power-law multiplier and
power coefficient. If not defined, a guess is made best on linear
fitting of the log-transformed `x` and `y` data.}
}
\value{
dataframe with the unadjusted R^2 value of best fit (field `r2`),
  the multiplier (`multiplier`) and power coefficient (`power`), the
  standard deviation of the residuals (`sd`) and the loglikelihood score
  (`loglikelihood`)
}
\description{
Fit a power law function between `x` and `y` in the form y = a*x^b.

The fit is generated using the `stats::nlm()` non-linear least-squares
regression function in R.
}
\examples{
x <- seq(2, 10, l = 25)
y <- 10*x^-0.5 * rweibull(length(x), shape = 4, scale = 1/gamma(1 + 1/4))
ft <- power_nls_fit(x, y)

xp <- seq(min(x), max(x), l = 101)
yp <- ft$multiplier*xp^ft$power

plot(x, y)
lines(xp, yp)

}
