% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_nls.R
\name{power_normal_sdpower_coef}
\alias{power_normal_sdpower_coef}
\title{Intermediate coefficients used in function `power_normal_sdpower_fit()`}
\usage{
power_normal_sdpower_coef(beta, delta, x, y, weights = rep(1, length(x)))
}
\arguments{
\item{beta}{power coefficient of the fit}

\item{delta}{power coefficient of the standard deviation of residuals}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a unitless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{weighting for each measurement. Default = 1, but a strong
case can be made for weighting with `weights = x^2` because of the
large effect of thick roots on root-reinforcement}
}
\value{
vector with intermediate coefficients
}
\description{
Defines eight intermediate variables that are used in the root solving
procedure used in the function `power_normal_sdpower_fit()` in order
to obtain fitting coefficients.
}
