% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_nls.R
\name{power_normal_sdpower_root}
\alias{power_normal_sdpower_root}
\title{Root function that is to be solved in `power_normal_sdpower_fit()`}
\usage{
power_normal_sdpower_root(par, x, y, weights = rep(1, length(x)))
}
\arguments{
\item{par}{two-parameter vector with values of beta and delta (power of
power law curve, power of standard deviation of residuals)}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a unitless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{weighting for each measurement. Default = 1, but a strong
case can be made for weighting with `weights = x^2` because of the
large effect of thick roots on root-reinforcement}
}
\value{
two-parameter vector that must be equal to zero at the correct
  values of `par`
}
\description{
Functions that need to be root-solved with respect to input arguments
`par`, as part of function `power_normal_sdpower_fit()`.
}
