% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_root}
\alias{power_root}
\title{Derivative of loglikelihood score for power law probability density}
\usage{
power_root(
  b,
  x,
  xmin = min(x),
  xmax = max(x),
  weights = rep(1, length(x)),
  eps = .Machine$double.eps^0.5
)
}
\arguments{
\item{b}{value for exponent}

\item{x}{vector with observations (must all be finite and >0)}

\item{xmin, xmax}{min and max limits of the density function}

\item{weights}{weighting for each observation in \code{x}}

\item{eps}{small value, to catch case where b == -1}
}
\value{
derivative of loglikelihood score
}
\description{
Returns the derivative of the loglikelihood function for a power law
probability distribution function. By solving this function == 0, the
power coefficient at which the loglikelihood score is maximised can be
found.
}
\keyword{internal}
