% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_weibull.R
\name{power_weibull_likelihood}
\alias{power_weibull_likelihood}
\title{Likelihood function for power-law weibull loglikelihood fitting}
\usage{
power_weibull_likelihood(par, lx, ly, weights = rep(1, length(lx)))
}
\arguments{
\item{par}{fitting parameter vector: log(multiplier), power, log(shape).}

\item{lx, ly}{vectors with log-transformed x and y measurements}

\item{weights}{weighting for each measurement}
}
\value{
negative log-likelihood score
}
\description{
Calculates the negative loglikelihood of the power-law + weibull shape
fit. Multiplier and weibull shape parameters are inputted as log-
transformed values to avoid issues with negative numbers.
}
