% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_weibull.R, R/power_weibull_OLD.R
\name{power_weibull_logdensity}
\alias{power_weibull_logdensity}
\title{Log-tranformed Weibull probability with power-law scale parameter}
\usage{
power_weibull_logdensity(x, y, alpha, power, shape)

power_weibull_logdensity(x, y, alpha, power, shape)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a unitless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{alpha}{vector with alpha-values}

\item{power}{vector with power-law power coefficients}

\item{shape}{vector with Weibull shape parameters}
}
\value{
vector with log-transformed probabilities for each x,y observation

vector with log-transformed probabilities for each x,y observation
}
\description{
Calculate the log-transformed Weibull probability, where the scale parameter
of the distribution follows a power-law curve with `x` in the shape

  y = alpha*x^power.

The alpha parameter is related to the multiplier of the best power-law
fit through the mean of the Weibull distribution, i.e.

  alpha = gamma(1 + 1/shape)*multiplier

Calculate the log-transformed Weibull probability, where the scale parameter
of the distribution follows a power-law curve with `x` in the shape

  y = alpha*x^power.

The alpha parameter is related to the multiplier of the best power-law
fit through the mean of the Weibull distribution, i.e.

  alpha = gamma(1 + 1/shape)*multiplier
}
\examples{
multiplier <- 10
power <- -0.5
shape <- 4
alpha <- multiplier/gamma(1 + 1/shape)
x <- seq(2, 8, l = 6)
y <- multiplier*x^power*rweibull(length(x), shape = shape, scale = 1/gamma(1 + 1/shape))
logp <- power_weibull_logdensity(x, y, alpha, power, shape)

p <- pweibull(y/(multiplier*x^power), shape = shape, scale = 1/gamma(1 + 1/shape))

exp(log(p))
p

multiplier <- 10
power <- -0.5
shape <- 4
alpha <- multiplier/gamma(1 + 1/shape)
x <- seq(2, 8, l = 6)
y <- multiplier*x^power*rweibull(length(x), shape = shape, scale = 1/gamma(1 + 1/shape))
logp <- power_weibull_logdensity(x, y, alpha, power, shape)

p <- pweibull(y/(multiplier*x^power), shape = shape, scale = 1/gamma(1 + 1/shape))

exp(log(p))
p

}
