% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_weibull.R, R/power_weibull_OLD.R
\name{power_weibull_plot}
\alias{power_weibull_plot}
\title{GGplot for power-law with Weibull fitting}
\usage{
power_weibull_plot(
  dr,
  y,
  multiplier = NULL,
  power = NULL,
  shape = NULL,
  confidence = c(0.5, 0.95),
  annotations = NULL,
  xlab = expression("Root diameter" ~ d[r] ~ "[mm]"),
  ylab = expression("Tensile strength" ~ t[r * "," * u] ~ "[MPa]"),
  xlim = c(0, NA),
  ylim = c(0, NA),
  ticks = 7,
  n = 101,
  dr0 = 1,
  settings = plot_settings()
)

power_weibull_plot(
  dr,
  y,
  multiplier = NULL,
  power = NULL,
  shape = NULL,
  confidence = c(0.5, 0.95),
  annotations = NULL,
  xlab = expression("Root diameter" ~ d[r] ~ "[mm]"),
  ylab = expression("Tensile strength" ~ t[r * "," * u] ~ "[MPa]"),
  xlim = c(0, NA),
  ylim = c(0, NA),
  ticks = 7,
  n = 101,
  dr0 = 1,
  settings = plot_settings()
)
}
\arguments{
\item{dr}{measured x-values (typically, root diameter)}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{multiplier}{power-law multiplier for best fit}

\item{power}{power-law power coefficient for best fit}

\item{shape}{Weibull shape parameter}

\item{confidence}{vector with prediction interval significance. For example
`confidence = c(0.5, 0.95)` will show the areas that contain 50% and 95%
of predicted measurements respectively.}

\item{annotations}{dataframe with labels to annotate, generated by function
`create_annotations()`}

\item{xlab, ylab}{x and y-axis labels}

\item{xlim, ylim}{two-element vectors with min and max values for the x and
y-axes. If axis limit can be freely chosen, specify the value of the min
and/or max as `NA`. The function will try to find nice axis limits for all
`NA` values using the function `round_range()`,}

\item{ticks}{maximum number of ticks on the axis. Used in function
`round_range()`}

\item{n}{number of (equally spaced) points along the best power-law fit}

\item{dr0}{reference diameter for power-law fits}

\item{settings}{list with default plot settings, as generated by the function
`plot_settings`}
}
\value{
ggplot object

ggplot object
}
\description{
Generate a plot, using ggplot, of a power-law fit using weibull
probabilities. Will also show the best fit and prediction intervals if
values are given as input.

Generate a plot, using ggplot, of a power-law fit using weibull
probabilities. Will also show the best fit and prediction intervals if
values are given as input.
}
\examples{
# input parameters
dr <- seq(2, 9.5, l = 50)
multiplier <- 10
power <- -0.3
shape <- 4

# generate some y-data
y <- multiplier*dr^power*rweibull(length(dr), shape = shape, scale = 1/gamma(1 + 1/shape))

# generate some labels
ann <- create_annotations(
  c(multiplier, power, shape),
  prefix = c("t[r*','*u*','*0]==", "beta[t]==", "kappa[t]=="),
  suffix = c("~'MPa'", "", ""),
  parse = TRUE
)

# plot
power_weibull_plot(dr, y, multiplier, power, shape, annotations = ann)

# input parameters
dr <- seq(2, 9.5, l = 50)
multiplier <- 10
power <- -0.3
shape <- 4

# generate some y-data
y <- multiplier*dr^power*rweibull(length(dr), shape = shape, scale = 1/gamma(1 + 1/shape))

# generate some labels
ann <- create_annotations(
  c(multiplier, power, shape),
  prefix = c("t[r*','*u*','*0]==", "beta[t]==", "kappa[t]=="),
  suffix = c("~'MPa'", "", ""),
  parse = TRUE
)

# plot
power_weibull_plot(dr, y, multiplier, power, shape, annotations = ann)

}
