% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw.R
\name{powerlaw_confidenceinterval}
\alias{powerlaw_confidenceinterval}
\title{Generate confidence interval for power-law fit}
\usage{
powerlaw_confidenceinterval(x, multiplier, exponent, Sigma, level = 0.95)
}
\arguments{
\item{x}{a-values at which to generate the interval}

\item{multiplier}{fitted power law multiplier}

\item{exponent}{fitted power law exponent}

\item{Sigma}{Variance-covariance matrix, with row and column names
indicating the variable names}

\item{level}{significance level}
}
\value{
dataframe with fields:
\itemize{
\item \code{x}: x-coordinates
\item \code{y}: average (power law)
\item \code{ymin}: lower limit of confidence interval
\item \code{ymax}: upper limit of confidence interval
}
}
\description{
Generate confidence intervals for power law fits, based on the
first delta method, and a calculate variance-covariance matrix
}
\examples{
# generate some data
y0 <- 20
beta <- -0.5
kappa <- 4
lambda <- 1/gamma(1 + 1/kappa)
x <- seq(1, 8, l = 101)
y <- y0*x^beta*rweibull(length(x), kappa, lambda)

# weibull fit
ftw <- powerlaw_fit(x, y, "weibull")
# generate covariance matrix
Sw <- powerlaw_covariancematrix(
  x, y, "weibull",
  multiplier = ftw$multiplier,
  exponent = ftw$exponent,
  shape = ftw$shape
)
Cw <- powerlaw_confidenceinterval(
  seq(min(x), max(x), l = 101),
  ftw$multiplier,
  ftw$exponent,
  Sw
)
# plot
plot(x, y)
lines(Cw$x, Cw$y, col = "red")
lines(Cw$x, Cw$ymin, col = "blue")
lines(Cw$x, Cw$ymax, col = "blue")

# gamma fit
ftg <- powerlaw_fit(x, y, "gamma")
# generate covariance matrix
Sg <- powerlaw_covariancematrix(
  x, y, "gamma",
  multiplier = ftg$multiplier,
  exponent = ftg$exponent,
  shape = ftg$shape
)
Cg <- powerlaw_confidenceinterval(
  seq(min(x), max(x), l = 101),
  ftw$multiplier,
  ftw$exponent,
  Sg
)
# plot
plot(x, y)
lines(Cg$x, Cg$y, col = "red")
lines(Cg$x, Cg$ymin, col = "blue")
lines(Cg$x, Cg$ymax, col = "blue")


}
