% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_gamma.R
\name{powerlaw_gamma_ks}
\alias{powerlaw_gamma_ks}
\title{Calculate Kolmogorov-Smirnov parameters for power-law fit + gamma}
\usage{
powerlaw_gamma_ks(x, y, multiplier, exponent, shape)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{multiplier, exponent}{multiplier and exponent for power-law fit
describing the mean}

\item{shape}{shape parameter for gamma distribution of residuals}
}
\value{
list with fields
\itemize{
\item \code{ks_distance}: Kolmogorov-Smirnov distance
}
}
\description{
Calculate Kolmogorov-Smirnov parameter for power-law fit with gamma
distributed residuals
}
\examples{
x <- seq(1, 7, l = 51)
y <- stats::rgamma(length(x), 4, scale = 2*1/4)
powerlaw_gamma_ks(x, y, 2, 0, 4)

}
