% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_gamma.R
\name{powerlaw_gamma_loglikelihood}
\alias{powerlaw_gamma_loglikelihood}
\title{Calculate power-law gamma loglikelihood}
\usage{
powerlaw_gamma_loglikelihood(par, x, y, weights = rep(1, length(x)), deriv = 0)
}
\arguments{
\item{par}{vector with fitting parameters (power-law multiplier, power-law
exponent, and gamma shape parameter)}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}

\item{deriv}{order of partial derivative requested}
}
\value{
loglikelihood, or its partial derivatives to order `deriv`
}
\description{
Calculate the weighted loglikelihood for a power-law fit with gamma
residuals. Can also be used to calculate the first or second partial
derivative with respect to fitting parameters.
}
\examples{
# parameters
x <- seq(1, 7, l = 101)
y0 <- 25
beta <- -0.5
k <- 4
y <- y0*x^beta*stats::rgamma(length(x), shape = k, scale = 1/k)
w <- stats::runif(length(x), 0.8, 1.2)

# test logliklihood
sum(w*stats::dgamma(y, k, scale = y0*x^beta/k, log = TRUE))
powerlaw_gamma_loglikelihood(c(y0, beta, k), x, y, weights = w)

# Test first derivative
eps <- 1e-9
par <- c(y0, beta, k)
f0 <- powerlaw_gamma_loglikelihood(par, x, y, weights = w)
f1 <- powerlaw_gamma_loglikelihood(par + c(eps, 0, 0), x, y, weights = w)
f2 <- powerlaw_gamma_loglikelihood(par + c(0, eps, 0), x, y, weights = w)
f3 <- powerlaw_gamma_loglikelihood(par + c(0, 0, eps), x, y, weights = w)
(c(f1, f2, f3) - f0)/eps
powerlaw_gamma_loglikelihood(par, x, y, weights = w, deriv = 1)

# test second derivative
f0 <- powerlaw_gamma_loglikelihood(par, x, y, weights = w, deriv = 1)
f1 <- powerlaw_gamma_loglikelihood(par + c(eps, 0, 0), x, y, weights = w, deriv = 1)
f2 <- powerlaw_gamma_loglikelihood(par + c(0, eps, 0), x, y, weights = w, deriv = 1)
f3 <- powerlaw_gamma_loglikelihood(par + c(0, 0, eps), x, y, weights = w, deriv = 1)
(cbind(f1, f2, f3) - f0)/eps
powerlaw_gamma_loglikelihood(par, x, y, weights = w, deriv = 2)

}
