% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_gamma.R
\name{powerlaw_gamma_root_k}
\alias{powerlaw_gamma_root_k}
\title{Root solve shape in gamma power-law}
\usage{
powerlaw_gamma_root_k(k, beta, x, y, weights = rep(1, length(x)))
}
\arguments{
\item{k}{unknown gamma distribution shape parameter}

\item{beta}{known power-law exponen}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
value of root function to solve
}
\description{
Root solve equation to obtain the Gamma distribution shape parameter
in power-law fitting with gamma-distribution
}
