% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_gumbel.R
\name{powerlaw_gumbel_ks}
\alias{powerlaw_gumbel_ks}
\title{Calculate Kolmogorov-Smirnov parameters for power-law fit + gumbel}
\usage{
powerlaw_gumbel_ks(
  x,
  y,
  multiplier,
  exponent,
  scale,
  gamma = 0.577215664901533
)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{multiplier, exponent}{multiplier and exponent for power-law fit
describing the mean}

\item{scale}{gumbel scale parameter}

\item{gamma}{Euler-Mascheroni constant}
}
\value{
list with fields
\itemize{
\item \code{ks_distance}: Kolmogorov-Smirnov distance
}
}
\description{
Calculate Kolmogorov-Smirnov parameter for power-law fit with gumbel
distributed residuals
}
\examples{
# generate some data
y0 <- 20
beta <- -0.5
theta0 <- 2
gamma <- 0.57721566490153286060651209008240243104215933593992
x <- seq(1, 6, l = 51)
y <- y0*x^beta*rgumbel(
  length(x),
  1 - gamma*theta0/y0,
  theta0/y0
)

# fit
ft <- powerlaw_gumbel_fit(x, y)

# ks distance
powerlaw_gumbel_ks(x, y, ft$multiplier, ft$exponent, ft$scale)

}
