% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_logistic.R
\name{powerlaw_logistic_ks}
\alias{powerlaw_logistic_ks}
\title{Calculate Kolmogorov-Smirnov parameters for power-law fit + logistic}
\usage{
powerlaw_logistic_ks(x, y, multiplier, exponent, scale)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{multiplier, exponent}{multiplier and exponent for power-law fit
describing the mean}

\item{scale}{scale parameter for standard deviation of residuals, at x = 1}
}
\value{
list with fields
\itemize{
\item \code{ks_distance}: Kolmogorov-Smirnov distance
}
}
\description{
Calculate Kolmogorov-Smirnov parameter for power-law fit with logistic
distributed residuals
}
\examples{
y0 <- 20
beta <- -0.5
s0 <- 4
x <- seq(1, 6, l = 51)
y <- abs(rlogis(length(x), y0*x^beta, s0*x^beta))
w <- runif(length(x), 0.8, 1.2)

ft <- powerlaw_logistic_fit(x, y, weights = w)

powerlaw_logistic_ks(x, y, ft$multiplier, ft$exponent, ft$scale)

}
