% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_logistic.R
\name{powerlaw_logistic_loglikelihood}
\alias{powerlaw_logistic_loglikelihood}
\title{Calculate power-law logistic loglikelihood}
\usage{
powerlaw_logistic_loglikelihood(
  par,
  x,
  y,
  weights = rep(1, length(x)),
  deriv = 0
)
}
\arguments{
\item{par}{vector with fitting parameters (power-law multiplier, power-law
exponent, and logistic shape parameter)}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}

\item{deriv}{order of partial derivative requested}
}
\value{
loglikelihood, or its partial derivatives to order `deriv`
}
\description{
Calculate the weighted loglikelihood for a power-law fit with logistic
residuals. Can also be used to calculate the first or second partial
derivative with respect to fitting parameters.
}
\examples{
# parameters and data
y0 <- 20
beta <- -0.5
s0 <- 4
x <- seq(1, 8, l = 100)
y <- stats::rlogis(length(x), y0*x^beta, s0*x^beta)
w <- runif(length(x), 0.8, 1.2)

# check loglikelihood function
par <- c(y0, beta, s0)
powerlaw_logistic_loglikelihood(par, x, y, weights = w, deriv = 0)
sum(w*stats::dlogis(y, y0*x^beta, s0*x^beta, log = TRUE))

# test first derivative
eps <- 1e-6
f0 <- powerlaw_logistic_loglikelihood(par, x, y, weights = w, deriv = 0)
f1 <- powerlaw_logistic_loglikelihood(par + c(eps, 0, 0), x, y, weights = w, deriv = 0)
f2 <- powerlaw_logistic_loglikelihood(par + c(0, eps, 0), x, y, weights = w, deriv = 0)
f3 <- powerlaw_logistic_loglikelihood(par + c(0, 0, eps), x, y, weights = w, deriv = 0)
(c(f1, f2, f3) - f0)/eps
powerlaw_logistic_loglikelihood(par, x, y, weights = w, deriv = 1)

# test second derivative
f0 <- powerlaw_logistic_loglikelihood(par, x, y, weights = w, deriv = 1)
f1 <- powerlaw_logistic_loglikelihood(par + c(eps, 0, 0), x, y, weights = w, deriv = 1)
f2 <- powerlaw_logistic_loglikelihood(par + c(0, eps, 0), x, y, weights = w, deriv = 1)
f3 <- powerlaw_logistic_loglikelihood(par + c(0, 0, eps), x, y, weights = w, deriv = 1)
(cbind(f1, f2, f3) - f0)/eps
powerlaw_logistic_loglikelihood(par, x, y, weights = w, deriv = 2)

}
