% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_lognormal.R
\name{powerlaw_lognormal_ks}
\alias{powerlaw_lognormal_ks}
\title{Calculate Kolmogorov-Smirnov parameters for power-law fit + lognormal}
\usage{
powerlaw_lognormal_ks(x, y, multiplier, exponent, sdlog)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{multiplier, exponent}{multiplier and exponent for power-law fit
describing the mean}

\item{sdlog}{standard deviation for for power-law fit
describing the log-standard deviation of residuals}
}
\value{
list with fields
\itemize{
\item \code{ks_distance}: Kolmogorov-Smirnov distance
}
}
\description{
Calculate Kolmogorov-Smirnov parameter for power-law fit with lognormal
distributed residuals
}
\examples{
y0 <- 20
beta <- -0.5
sdlog <- 0.3
muL <- log(y0) - 0.5*sdlog^2
x <- seq(1, 8, l = 101)
y <- x^beta*rlnorm(length(x), muL, sdlog)

ft <- powerlaw_lognormal_fit(x, y)

powerlaw_lognormal_ks(x, y, ft$multiplier, ft$exponent, ft$sdlog)

}
