% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freebeta_initialguess}
\alias{powerlaw_normal_freebeta_initialguess}
\title{Generate initial guess for `powerlaw_normal_fit()` with known delta}
\usage{
powerlaw_normal_freebeta_initialguess(
  x,
  y,
  delta = 0,
  weights = rep(1, length(x))
)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{delta}{known value for standard deviation power law exponent}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
estimate for power-law exponents for the mean
}
\description{
Generate an initial guess for mean power-law exponen, to be used in
function `powerlaw_normal_fit()`. The standard deviation power-law exponent
(delta) is assumed to be known.

The mean power-law exponent (beta) is estimated using linear regression on
log-transformed x and y data.
}
\keyword{internal}
