% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freebetadelta_initialguess}
\alias{powerlaw_normal_freebetadelta_initialguess}
\title{Generate initial guess for `powerlaw_normal_fit()` with two unknown exponents}
\usage{
powerlaw_normal_freebetadelta_initialguess(
  x,
  y,
  weights = rep(1, length(x)),
  method = "bisection",
  range = c(-1, 1),
  sd_exponent_range = c(-2, 2),
  n_range = 9
)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}

\item{method}{Only used when finding a single root.
Choose `newton` for gradient descent solving, using the
`rootSolve::multiroot()` function, or `bisection` for bisection root
solving algorithm using the `stats::uniroot()` function.
When using a multivariate root, `rootSolve::multiroot()` will always be
used.}

\item{range}{two-value array to add to best guess, to define initial
interval for bisection algorithm}

\item{sd_exponent_range}{search range around beta, for initial guess of
delta in the case where both beta and delta are to be fitted. See function
`powerlaw_normal_freebetadelta_initialguess()` for more details.}

\item{n_range}{number of guesses for delta to be tried on range
`sd_exponent_range`. See function
`powerlaw_normal_freebetadelta_initialguess()` for more details.}
}
\value{
estimate for power-law exponents for the mean and standard deviation
}
\description{
Generate an initial guess for mean and standard deviation power-law
exponents, to be used in function `powerlaw_normal_fit()`.

The mean power-law exponent (beta) is estimated using linear regression on
log-transformed x and y data.

Subsequently, `n` equally spaced guess for the standard deviation
power-law exponents are tried, on the domain
(beta_lm + `sd_exponent_range`). The beta and delta value of the fit
with the lowest loglikelihood are taken as the initial guess
}
\keyword{internal}
