% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freebetadelta_loglikelihood}
\alias{powerlaw_normal_freebetadelta_loglikelihood}
\title{Calculate power-law normal loglikelihood - separate power law for sd}
\usage{
powerlaw_normal_freebetadelta_loglikelihood(
  par,
  x,
  y,
  weights = rep(1, length(x)),
  deriv = 0
)
}
\arguments{
\item{par}{vector with fitting parameters (power-law multiplier, power-law
exponent, standard deviation power-law multiplier, and standard deviation
power-law exponent)}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}

\item{deriv}{order of partial derivative requested}
}
\value{
loglikelihood, or its partial derivatives to order `deriv`
}
\description{
Calculate the weighted loglikelihood for a power-law fit with normal
residuals. Standard deviation follows its own power-law.
Can also be used to calculate the first or second partial
derivative with respect to fitting parameters.
}
\keyword{internal}
