% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freebetadelta_root}
\alias{powerlaw_normal_freebetadelta_root}
\title{Root solve exponent in normal power-law - separate power law for sd}
\usage{
powerlaw_normal_freebetadelta_root(par, x, y, weights = rep(1, length(x)))
}
\arguments{
\item{par}{vector with mean and standard deviation power-law exponents}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
value of root functions to solve
}
\description{
Root solve equation to obtain the exponent in power-law fitting
with normal distribution with separate power-law describing the standard
deviation
}
