% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freedelta_initialguess}
\alias{powerlaw_normal_freedelta_initialguess}
\title{Generate initial guess for `powerlaw_normal_fit()` with known beta}
\usage{
powerlaw_normal_freedelta_initialguess(
  x,
  y,
  beta = 0,
  weights = rep(1, length(x))
)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{beta}{known value for power law exponent}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
estimate for power-law exponents for the standard deviation
}
\description{
Generate an initial guess for standard deviation mean power-law exponent,
to be used in function `powerlaw_normal_fit()`. The mean power-law exponent
(beta) is assumed to be known.

Currently, a very simple guess delta = beta is made.
}
\keyword{internal}
