% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freedelta_loglikelihood}
\alias{powerlaw_normal_freedelta_loglikelihood}
\title{Calculate power-law normal loglikelihood - fixed beta}
\usage{
powerlaw_normal_freedelta_loglikelihood(
  par,
  x,
  y,
  beta = 0,
  weights = rep(1, length(x)),
  deriv = 0
)
}
\arguments{
\item{par}{vector with fitting parameters (power-law multiplier,
standard deviation power-law multiplier, standard deviation power-law
exponent)}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{beta}{known power-law exponent for the mean}

\item{weights}{vector with weighting for each observation (default: 1)}

\item{deriv}{order of partial derivative requested}
}
\value{
loglikelihood, or its partial derivatives to order `deriv`
}
\description{
Calculate the weighted loglikelihood for a power-law fit with normal
residuals. The power-law exponent for the mean (beta) is
assumed to be known.

Can also be used to calculate the first or second partial
derivative with respect to fitting parameters.
}
\keyword{internal}
