% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freedelta_root}
\alias{powerlaw_normal_freedelta_root}
\title{Root solve exponent in normal power-law - known beta}
\usage{
powerlaw_normal_freedelta_root(
  delta,
  x,
  y,
  beta = 0,
  weights = rep(1, length(x))
)
}
\arguments{
\item{delta}{value for standard deviation power-law exponent}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{beta}{known power-law exponent for the mean}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
value of root function to solve
}
\description{
Root solve equation to obtain the standard deviation exponent in power-law
fitting with normal distribution. Power-law exponent for mean (beta)
is assumed to be known.
}
\keyword{internal}
