% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freedelta_root_jacobian}
\alias{powerlaw_normal_freedelta_root_jacobian}
\title{Jacobian of function `powerlaw_normal_freedelta_root()`}
\usage{
powerlaw_normal_freedelta_root_jacobian(
  delta,
  x,
  y,
  beta = 0,
  weights = rep(1, length(x))
)
}
\arguments{
\item{delta}{value for standard deviation power-law exponent}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{beta}{known power-law exponent for the mean}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
derivative of `powerlaw_normal_freedelta_root()` with respect to
  `delta`#'
}
\description{
Returns the derivative of the function `powerlaw_normal_freedelta_root()`
with respect to input argument `delta`
}
\keyword{internal}
