% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_ks}
\alias{powerlaw_normal_ks}
\title{Calculate Kolmogorov-Smirnov parameters for power-law fit + normal}
\usage{
powerlaw_normal_ks(x, y, multiplier, exponent, sd_multiplier, sd_exponent = 0)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{multiplier, exponent}{multiplier and exponent for power-law fit
describing the mean}

\item{sd_multiplier, sd_exponent}{multiplier and exponent for power-law fit
describing the standard deviation of residuals}
}
\value{
list with fields
\itemize{
\item \code{ks_distance}: Kolmogorov-Smirnov distance
}
}
\description{
Calculate Kolmogorov-Smirnov parameter for power-law fit with normally
distributed residuals
}
\examples{
x <- seq(1, 7, l = 251)
y <- stats::rnorm(length(x), 2, 0.5)
powerlaw_normal_ks(x, y, 2, 0, 0.5)

}
