% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_linkedbetadelta_root}
\alias{powerlaw_normal_linkedbetadelta_root}
\title{Root solve exponent in normal power-law - beta=delta}
\usage{
powerlaw_normal_linkedbetadelta_root(beta, x, y, weights = rep(1, length(x)))
}
\arguments{
\item{beta}{value for mean power-law exponent}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
value of root function to solve
}
\description{
Root solve equation to obtain the mean exponent in power-law
fitting with normal distribution (beta). Power-law exponent for the
standard deviation (delta) is assumed to be equal to beta.
}
\keyword{internal}
