% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_uniform.R
\name{powerlaw_uniform_ks}
\alias{powerlaw_uniform_ks}
\title{Calculate Kolmogorov-Smirnov parameters for power-law fit + uniform}
\usage{
powerlaw_uniform_ks(x, y, multiplier, exponent, width)
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{multiplier, exponent}{multiplier and exponent for power-law fit
describing the mean}

\item{width}{with of the distribution, at x = 1}
}
\value{
list with fields
\itemize{
\item \code{ks_distance}: Kolmogorov-Smirnov distance
}
}
\description{
Calculate Kolmogorov-Smirnov parameter for power-law fit with uniformly
distributed residuals
}
\examples{
y0 <- 20
beta <- -0.5
range <- 10
x <- seq(1, 8, l = 51)
y <- stats::runif(length(x), (y0 - 0.5*range)*x^beta, (y0 + 0.5*range)*x^beta)

ft <- powerlaw_uniform_fit(x, y)

powerlaw_uniform_ks(x, y, ft$multiplier, ft$exponent, ft$width)

}
