% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_weibull.R
\name{powerlaw_weibull_initialguess}
\alias{powerlaw_weibull_initialguess}
\title{Generate initial guess for `powerlaw_weibull_fit()`}
\usage{
powerlaw_weibull_initialguess(x, y, weights = rep(1, length(x)))
}
\arguments{
\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
estimate for power-law exponent and weibull shape parameter
}
\description{
Generate an intiial guess for power-law exponent and weibull shape parameter,
to be used in function `powerlaw_weibull_fit()`.

The power-law exponent is estimated using linear regression on
log-transformed x and y data. The weibull shape parameter is subsequently
estimated from weibull fitting of the scaled data (y/x^beta)
}
\keyword{internal}
