% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_weibull.R
\name{powerlaw_weibull_root}
\alias{powerlaw_weibull_root}
\title{Root to solve for power-law Weibull fitting}
\usage{
powerlaw_weibull_root(par, x, y, weights = rep(1, length(x)))
}
\arguments{
\item{par}{fitting parameter (power-law exponent, and Weibull shape
parameter)}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
two-parameter vector with roots
}
\description{
Root equation to solve for power-law Weibull fitting
}
