% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{rgumbel}
\alias{rgumbel}
\title{Generate random Gumbel values}
\usage{
rgumbel(n, location = 0, scale = 1)
}
\arguments{
\item{n}{number of samples to draw}

\item{location}{Gumbel location parameter}

\item{scale}{Gumbel scale parameter}
}
\value{
vector with samples
}
\description{
Draw random values from a Gumbel probability distribution
}
\examples{
hist(rgumbel(1000, location = 20, scale = 2))

}
