% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootclasscount.R
\name{rootclass_constraints}
\alias{rootclass_constraints}
\title{Generate contraints for `rootclass_fit()` optimalisation}
\usage{
rootclass_constraints(ns, xmin, xmax, y, fixed = rep(NA, 2 * ns + 1))
}
\arguments{
\item{ns}{number of power-law segments to use in fit}

\item{xmin, xmax}{vectors with lower and upper limits for each root diameter
class. `xmax > xmin >= TRUE` for each element.}

\item{y}{vector with number of roots (or total root length) per class}

\item{fixed}{vector with `2*ns + 1` elements, with values of the
fitting parameter (`ns + 1` breakpoints, `ns` power-law coefficients)
that should be fixed and therefore not fitted. Values that are `NA` will
be fitted while any non-NA will be held constant}
}
\value{
a list with the constraint matrix (`ui`) and vector (`ci`)
}
\description{
Generate constraint matrix `ui` and constraint vector `ci` to use in
the fitting function `rootclass_fit()`, which uses the `stats::constrOptim()`
optimalisation method (see `stats::constrOptim()` for more detail).

The following constraints are used:
- Smallest breakpoint larger than 0
- Smallest breakpoint smaller than the smallest upper class limit in `xmax`
  that contains roots (`y > 0`)
- Breakpoints vector in increasing order
- Largest breakpoint larger than the largest lower class limit in `xmin`
  that contains roots (`y > 0`)
}
\examples{
ns <- 3
xmin <- seq(2, 5)
xmax <- seq(3, 6)
y <- c(1, 4)
rootclass_constraints(ns, xmin, xmax, y)

}
