% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootclasscount.R
\name{rootclass_cumulative_observations}
\alias{rootclass_cumulative_observations}
\title{Cumulative probability density curve for root class count observations}
\usage{
rootclass_cumulative_observations(xmin, xmax, y)
}
\arguments{
\item{xmin, xmax}{vectors with lower and upper limits for each root diameter
class. `xmax > xmin >= TRUE` for each element.}

\item{y}{vector with number of roots (or total root length) per class}
}
\value{
a dataframe with x-values (`x`) and the cumulative probability
  density (`y`)
}
\description{
Generate a observation-value versus cumulative probability density
curve for a series of root counts and root diameter classes
}
\examples{
no <- 9
xc <- seq(2, 7, l = no + 1)
xmin <- xc[1:no]
xmax <- xc[2:(no + 1)]
y <- (8 - (0.5*(xmin + xmax) - 5)^2)

rootclass_cumulative_observations(xmin, xmax, y)

}
