% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootclasscount.R
\name{rootclass_cumulative_plot}
\alias{rootclass_cumulative_plot}
\title{Plot observed and fitted cumulative density of root class distributions}
\usage{
rootclass_cumulative_plot(
  par,
  xmin,
  xmax,
  y,
  n = 101,
  xlab = expression("Root diameter" ~ d[r] ~ "[mm]"),
  ylab = "Cumulative probability density [-]",
  xlim = c(0, NA),
  ylim = c(0, 1),
  ticks = 7,
  settings = plot_settings()
)
}
\arguments{
\item{par}{dataframe with fitting values per segment. For more information,
see documentation in output argument in function `rootclass_fit()`}

\item{xmin, xmax}{vector with lower and upper limits of diameter classes}

\item{y}{vector with number of roots (or root length) in each class}

\item{n}{number of points to use for fitting line in each segment}

\item{xlab, ylab}{x and y-axis labels}

\item{xlim, ylim}{two-element vectors with min and max values for the x and
y-axes. If axis limit can be freely chosen, specify the value of the min
and/or max as `NA`. The function will try to find nice axis limits for all
`NA` values using the function `round_range()`,}

\item{ticks}{maximum number of ticks on the axis. Used in function
`round_range()`}

\item{settings}{list with default plot settings, as generated by the function
`plot_settings`}
}
\value{
ggplot object
}
\description{
Plot the observed and multisegment power-law fit for the cumulative density
distribution of variable `x`, for root diameter class count data

Plots are generated using the ggplot2 package.
}
\examples{
no <- 9
xc <- seq(2, 6.8, l = no + 1)
xmin <- xc[1:no]
xmax <- xc[2:(no + 1)]
y <- (8 - (0.5*(xmin + xmax) - 5)^2)

ns <- 2
par <- rootclass_fit(xmin, xmax, y, ns)$par

rootclass_cumulative_plot(par, xmin, xmax, y)

}
