% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootclasscount.R
\name{rootclass_loglikelihood}
\alias{rootclass_loglikelihood}
\title{Loglikelihood score of multi-segment power-law fit for root count data}
\usage{
rootclass_loglikelihood(
  par,
  xmin,
  xmax,
  y,
  fixed = rep(NA, length(par)),
  weights_multiplier = rep(1, length(xmin)),
  weights_power = rep(0, length(xmin))
)
}
\arguments{
\item{par}{vector with fitting parameters. For `ns` segments, the vector will
consist of `ns + 1` consecutive x-breakpoints, followed by `ns` power-law
cower coefficients}

\item{xmin, xmax}{vectors with lower and upper limits for each root diameter
class. `xmax > xmin >= TRUE` for each element.}

\item{y}{vector with number of roots (or total root length) per class}

\item{fixed}{vector with `2*ns + 1` elements, with values of the
fitting parameter (`ns + 1` breakpoints, `ns` power-law coefficients)
that should be fixed and therefore not fitted. Values that are `NA` will
be fitted while any non-NA will be held constant}

\item{weights_multiplier}{vector with fitting multiplication weights for
each class}

\item{weights_power}{vector with fitting power-law coefficients for each
class}
}
\value{
loglikelihood value (scalar)
}
\description{
Loglikelihood score of multi-segment power-law fit for root count data
}
\examples{
xb <- c(1.5, 3.5, 6.5, 7.5, 8.5)
b <- seq(-2, 2, l = length(xb) - 1)
par <- c(xb, b)
xmin <- c(2,3,4)
xmax <- c(2.5,4,6)
y <- c(3, 4, 5)

rootclass_loglikelihood(par, xmin, xmax, y)

}
