% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootclasscount.R
\name{rootclass_multipliers}
\alias{rootclass_multipliers}
\title{Scaling multipliers for power-laws for each segment}
\usage{
rootclass_multipliers(xb, b)
}
\arguments{
\item{xb}{vector with x-breakpoints (length = number of segments + 1)}

\item{b}{vector with power law coefficients for each segment}
}
\value{
a vector with power-law multipliers for each fitted segment
}
\description{
Calculates the power-law multiplication factors for each segment. These
are yet unscaled (so the total probability != 1), and instead assume
that the multiplier for the first segment is equal to 1.

Correct probability scaling is done within the function
`rootclass_loglikelihood()` instead.
}
\examples{
# check if correct, by checking if probability curve is C^0 continuous
xb <- c(1.5, 3.5, 6.5, 7.5, 8.5)
b <- seq(-2, 2, l = length(xb) - 1)
a <- rootclass_multipliers(xb, b)
df <- data.frame(a = a, b = b, x1 = xb[1:(length(xb) - 1)],
  x2 = xb[2:length(xb)])
df2 <- expand_grid_df(data.frame(s = seq(0, 1, l = 25)), df)
df2$x <- with(df2, x1 + s*(x2 - x1))
df2$y <- with(df2, a*x^b)
plot(df2$x, df2$y, "l")

}
