% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootclasscount.R
\name{rootclass_probfit}
\alias{rootclass_probfit}
\title{Total probability in fitting segment}
\usage{
rootclass_probfit(xb, a, b)
}
\arguments{
\item{xb}{vector with x-breakpoints (length = number of segments + 1)}

\item{a}{power-law multipliers for each segment, as calculated with the
function `rootclass_multipliers()`}

\item{b}{vector with power law coefficients for each segment}
}
\value{
vector with (unscaled) total probabilities
}
\description{
Calculate the total probability within each fitted power-law segment.
These values are not yet scaled for the total probability.
}
\examples{
xb <- c(1.5, 3.5, 6.5, 7.5, 8.5)
a <- seq(2, 4, l = length(xb) - 1)
b <- seq(-2, 2, l = length(xb) - 1)
rootclass_probfit(xb, a, b)

}
