% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootclasscount.R
\name{rootclass_probobs}
\alias{rootclass_probobs}
\title{Total probability in each measured root class}
\usage{
rootclass_probobs(
  xb,
  a,
  b,
  xmin,
  xmax,
  weights_multiplier = rep(1, length(xmin)),
  weights_power = rep(0, length(xmin))
)
}
\arguments{
\item{xb}{vector with x-breakpoints (length = number of segments + 1)}

\item{a}{power-law multipliers for each segment, as calculated with the
function `rootclass_multipliers()`}

\item{b}{vector with power law coefficients for each segment}

\item{xmin, xmax}{vectors with lower and upper limits for each root diameter
class. `xmax > xmin >= TRUE` for each element.}

\item{weights_multiplier}{vector with fitting multiplication weights for
each class}

\item{weights_power}{vector with fitting power-law coefficients for each
class}
}
\value{
vector with (unscaled) total probabilities
}
\description{
Calculate the total probability within each measured root class, based on
the fitted power-law segment. These values are not yet scaled for the total
probability.
}
\examples{
xb <- c(1.5, 3.5, 6.5, 7.5, 8.5)
a <- seq(2, 4, l = length(xb) - 1)
b <- seq(-2, 2, l = length(xb) - 1)
xmin <- c(2, 3, 4)
xmax <- c(2.5, 4, 20)
rootclass_probobs(xb, a, b, xmin, xmax)

}
