% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_constructpar}
\alias{rootcount_constructpar}
\title{Combine fitting parameters and known parameters into single vector}
\usage{
rootcount_constructpar(par, fixed = rep(NA, length(par)))
}
\arguments{
\item{par}{vector with fitting parameters}

\item{fixed}{vector with known parameters. `NA` values in this vector
indicate unknwon values (i.e., those in `par`).}
}
\value{
vector with all coefficients
}
\description{
Generate a single vector with fitting parameters and known ('fixed')
parameters for root count fitting.
}
\examples{
rootcount_construct_par(c(1,2,3), fixed = c(NA, 5, NA, NA, 6))

}
