% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_cumulative_observations}
\alias{rootcount_cumulative_observations}
\title{Generate a cumulative probability density curve for observations}
\usage{
rootcount_cumulative_observations(x)
}
\arguments{
\item{x}{vector with observations}
}
\value{
a dataframe with x-values (`x`) and the cumulative probability
density (`y`)
}
\description{
Generate a observation-value versus cumulative probability density
curve for a series of observations
}
\examples{
rootcount_cumulative_observations(seq(2, 5, l = 10))

}
