% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_cumulative_plot}
\alias{rootcount_cumulative_plot}
\title{Plot observed and fitted cumulative density of root distributions}
\usage{
rootcount_cumulative_plot(
  x,
  par,
  n = 101,
  xlab = expression("Root diameter" ~ d[r] ~ "[mm]"),
  ylab = "Cumulative probability density [-]",
  xlim = c(0, NA),
  ylim = c(0, 1),
  ticks = 7,
  settings = plot_settings()
)
}
\arguments{
\item{x}{vector with observations}

\item{par}{dataframe with fitting values per segment. For more information,
see documentation in output argument in function `rootcount_fit()`}

\item{n}{number of points to use for fitting line in each segment}

\item{xlab, ylab}{x and y-axis labels}

\item{xlim, ylim}{two-element vectors with min and max values for the x and
y-axes. If axis limit can be freely chosen, specify the value of the min
and/or max as `NA`. The function will try to find nice axis limits for all
`NA` values using the function `round_range()`,}

\item{ticks}{maximum number of ticks on the axis. Used in function
`round_range()`}

\item{settings}{list with default plot settings, as generated by the function
`plot_settings`}
}
\value{
ggplot object
}
\description{
Plot the observed and multisegment power-law fit for the cumulative density
distribution of variable `x`.

Plots are generated using the ggplot2 package.
}
\examples{
x <- c(
  rweibull(25, shape = 4, scale = 1),
  rweibull(25, shape = 6, scale = 3),
  rweibull(25, shape = 12, scale = 6)
)

ft <- rootcount_fit(x, 6)

rootcount_cumulative_plot(x, ft$par)
}
