% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_cumulative_prediction}
\alias{rootcount_cumulative_prediction}
\title{Predict fitted cumulative probability density for observations}
\usage{
rootcount_cumulative_prediction(x, par)
}
\arguments{
\item{x}{array with observations}

\item{par}{dataframe with fitting parameters per power-law segment. This
is the dataframe returned by the function `rootcount_fit` as the field
`par` in the outputted list}
}
\value{
array with cumulative probability densities
}
\description{
Calculate the fitted cumulative probability density for observations `x`,
given the fit results `par` generated by the function `rootcount_fit()`
or `rootclass_fit()`.
}
\examples{
x <- rweibull(100, 4, 6)
par <- rootcount_fit(x, 2)$par

xp <- seq(min(x), max(x), l = 251)
yp <- rootcount_cumulative_prediction(xp, par)

plot(xp, yp, "l")

}
