% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_density_fitted}
\alias{rootcount_density_fitted}
\title{Generate a probability density curve for multisegment fit}
\usage{
rootcount_density_fitted(dfp, n = 101)
}
\arguments{
\item{dfp}{dataframe with fitting parameters per power-law segment. This
is the dataframe returned by the function `rootcount_fit` as the field
`par` in the outputted list}

\item{n}{number of points to use on curve for each segment}
}
\value{
a dataframe with x-values (`x`), the cumulative probability
density (`y`), and an identifier for each bundle (`bundle_id`, integers
1, 2, 3 etc)
}
\description{
Generate a observation versus probability density
curve for a multisegment power-law fit, as generated by the
fitting function `rootcount_fit()`.
}
\examples{
ft <- rootcount_fit(seq(2, 6, l = 51), 2)
df <- rootcount_density_fitted(ft$par, n = 101)
plot(df$x, df$y, "l")

}
