% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggplot.R
\name{rootcount_density_plot}
\alias{rootcount_density_plot}
\title{Plot observed and fitted root probability distributions}
\usage{
rootcount_density_plot(
  x,
  df,
  bins = 20,
  n = 101,
  xlab = expression("Root diameter" ~ d[r] ~ "[mm]"),
  ylab = "Probability density [-]",
  xlim = c(0, NA),
  ylim = c(0, NA),
  ticks = 7,
  settings = plot_settings()
)
}
\arguments{
\item{x}{vector with observations}

\item{df}{dataframe with fitting values per segment. For more information,
see documentation in output argument in function `rootcount_fit()`}

\item{bins}{number of equally spaced bins to use to plot observed data}

\item{n}{number of points to use for fitting line in each segment}

\item{xlab, ylab}{x and y-axis labels}

\item{xlim, ylim}{two-element vectors with min and max values for the x and
y-axes. If axis limit can be freely chosen, specify the value of the min
and/or max as `NA`. The function will try to find nice axis limits for all
`NA` values using the function `round_range()`,}

\item{ticks}{maximum number of ticks on the axis. Used in function
`round_range()`}

\item{settings}{list with default plot settings, as generated by the function
`plot_settings`}
}
\value{
ggplot object
}
\description{
Plot the observed and multisegment power-law fit for the probability density
distribution of variable `x`. The observed data is binned using `bins`
number of equal-width bins, and per bin the average probabilty density is
calculated.

Plots are generated using the ggplot2 package.
}
\examples{
x <- c(
  rweibull(50, shape = 4, scale = 1),
  rweibull(50, shape = 6, scale = 3),
  rweibull(50, shape = 12, scale = 6)
)

ft <- rootcount_fit(x, 6)

rootcount_density_plot(x, ft$par, bins = 30)
rootcount_cumulative_plot(x, ft$par)
}
