% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_initialguess}
\alias{rootcount_initialguess}
\title{Initial guess for multisegment power-law fitting}
\usage{
rootcount_initialguess(x, ns)
}
\arguments{
\item{x}{vector with observations}

\item{ns}{number of segments to use}
}
\value{
a vector with an initial guess
}
\description{
Generates a single initial guess for multisegment power-law
fitting. The breakpoints are equally distributed between `min(x)` and
`max(x)`, while all power-law coefficients are assumed zero
}
\examples{
rootcount_initialguess(seq(1.2, 5.6, l = 50), 3)

}
