% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_initialguess_single}
\alias{rootcount_initialguess_single}
\title{Initial guess for single-segment power-law fitting}
\usage{
rootcount_initialguess_single(x)
}
\arguments{
\item{x}{vector with observations}
}
\value{
a vector with an initial guess
}
\description{
Generates an initial guess for single segment power-law
fitting. The only fitting parameter is the power-law power
coefficient
}
\examples{
rootcount_initialguess_single(seq(1.2, 5.6, l = 50))

}
