% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_loglikelihood}
\alias{rootcount_loglikelihood}
\title{Loglikelihood score of multi-segment power-law fit}
\usage{
rootcount_loglikelihood(
  par,
  x,
  xmin = min(x, na.rm = TRUE),
  xmax = max(x, na.rm = TRUE),
  weights = rep(1, length(x))
)
}
\arguments{
\item{par}{vector with fitting parameters. For n segments, the vector will
consist of n-1 consecutive x-breakpoints, followed by n power-law
cower coefficients}

\item{x}{vector with observations}

\item{xmin, xmax}{minimum and maximum values of `x` to use in the fit.
By default, `xmin = min(x)` and `xmax = max(x)` since this will maximise
the probability. These values may however be overwritten if required.}

\item{weights}{vector with weights for each observation. The individual
probabilities are raised to the power `weights`. By default, all
observations are weighted equally (`weights = 1`)}
}
\value{
scalar with weighted loglikelihood value
}
\description{
Return the (weighted) loglikelihood score for a multi-segment
power-law fit.
}
\examples{
# parameters
x <- seq(1.5, 6.5, l = 25)
xb <- c(2, 3.2, 5)
b <- c(1.1, 3, -2, -0.4)

# loglikelihood
rootcount_loglikelihood(c(xb, b), x)

}
