% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_loglikelihood_jacobian}
\alias{rootcount_loglikelihood_jacobian}
\title{Derivative of function `rootcount_loglikelihood()`}
\usage{
rootcount_loglikelihood_jacobian(
  par,
  x,
  xmin = min(x, na.rm = TRUE),
  xmax = max(x, na.rm = TRUE),
  weights = rep(1, length(x))
)
}
\arguments{
\item{par}{vector with fitting parameters. For n segments, the vector will
consist of n-1 consecutive x-breakpoints, followed by n power-law
cower coefficients}

\item{x}{vector with observations}

\item{xmin, xmax}{minimum and maximum values of `x` to use in the fit.
By default, `xmin = min(x)` and `xmax = max(x)` since this will maximise
the probability. These values may however be overwritten if required.}

\item{weights}{vector with weights for each observation. The individual
probabilities are raised to the power `weights`. By default, all
observations are weighted equally (`weights = 1`)}
}
\value{
a scalar containing the derivatives for each value in fitting vector
  `par`
}
\description{
Generates the derivative of the results of the function
`rootcount_loglikelihood()` with respect to fitting value input
`par`.

Function is vectorised.
}
\examples{
# Test jacobian by comparing to numberical solution
x <- seq(1.4, 6.9, l = 25)
xb <- c(2, 3.2, 5.1)
b <- c(1.1, 2.6, -2.1, -0.4)
par <- c(xb, b)
w <- seq(0.8, 1.2, l = length(x))

L <- rootcount_loglikelihood(par, x, weights = w)
J <- rootcount_loglikelihood_jacobian(par, x, weights = w)

eps <- 1e-6
J2 <- rep(0, length(par))
for (i in 1:length(par)) {
  dx <- rep(0, length(par))
  dx[i] <- eps
  J2[i] <- (rootcount_loglikelihood(par + dx, x, weights = w) - L)/eps
}

J
J2

}
