% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_loglikelihood_single}
\alias{rootcount_loglikelihood_single}
\title{Loglikelihood for fitting a single segment}
\usage{
rootcount_loglikelihood_single(
  par,
  x,
  xmin = min(x, na.rm = TRUE),
  xmax = max(x, na.rm = TRUE),
  weights = rep(1, length(x))
)
}
\arguments{
\item{par}{vector with fitting parameters. For n segments, the vector will
consist of n-1 consecutive x-breakpoints, followed by n power-law
cower coefficients}

\item{x}{vector with observations}

\item{xmin, xmax}{minimum and maximum values of `x` to use in the fit.
By default, `xmin = min(x)` and `xmax = max(x)` since this will maximise
the probability. These values may however be overwritten if required.}

\item{weights}{vector with weights for each observation. The individual
probabilities are raised to the power `weights`. By default, all
observations are weighted equally (`weights = 1`)}
}
\value{
scalar with weighted loglikelihood value
}
\description{
Calculate the loglikelihood score for a single-segment power-law
fit. In this case, the only fitting parameter is a single power-law power
coefficient.
}
\examples{
rootcount_loglikelihood_single(2, seq(2, 4, l = 10))

}
