% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_logprob}
\alias{rootcount_logprob}
\title{log-transformed probability for multi-segment power-law}
\usage{
rootcount_logprob(logxb, b, logx)
}
\arguments{
\item{logxb}{vector with log-transformed values of the x-breakpoints between
segments}

\item{b}{vector with power-law power coefficients for each segment}

\item{logx}{vector with log-transformed observations x}
}
\value{
vector with scaled log-transformed probabilities for each
  observation in `x`
}
\description{
Returns log-transformed probabilities for observations x. Probabilities are
scaled by an unknown multiplier, which is seperately determined (function
`rootcount_multiplier()`) based on the fact that the total probability
must equal one.
}
\examples{
x <- seq(1, 4, l = 14)
b <- c(1, 2, 3)
xb <- c(2.2, 3.6)
rootcount_logprob(log(xb), b, log(x))

}
