% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_multiplier}
\alias{rootcount_multiplier}
\title{Scaling multipliers for individual probabilities}
\usage{
rootcount_multiplier(xb, b, xmin, xmax)
}
\arguments{
\item{xb}{vector with x-breakpoints between segments}

\item{b}{vector with power-law power coefficients for each segment}

\item{xmin, xmax}{minimum and maximum values of `x` to use in the fit.
By default, `xmin = min(x)` and `xmax = max(x)` since this will maximise
the probability. These values may however be overwritten if required.}
}
\value{
vector with multiplier contributions of each segment, which, when
  summed give the probability multiplier
}
\description{
Calculates the scaling factor that should be applied to individual
probabilities to ensure that the total probability equals one.

Returns a vector, which, when summed, gives the value of the multiplier.

Probabilities  calculated using `rootcount_logprob()` should be divided
by this multiplier to get the actual probability.
}
\examples{
# TEST - does probability sum to one?
# parameters
x <- seq(1.5, 6.5, l = 1000)
xb <- c(2, 3.2, 5)
b <- c(1.1, 3, -2, -0.4)
xmin <- min(x)
xmax <- max(x)

# individual probabilities
pip <- exp(rootcount_logprob(log(xb), b, log(x)))
# integral prod
pt <- sum(rootcount_multiplier(xb, b, xmin, xmax))
# quick estmation of integral
sum(pip/pt)*diff(x)[1]

}
