% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_random}
\alias{rootcount_random}
\title{Generate random root diameters based on multisegment power-law probabilities}
\usage{
rootcount_random(n, par)
}
\arguments{
\item{n}{number of random observations required}

\item{par}{dataframe with fitting parameters per power-law segment. This
is the dataframe returned by the function `rootcount_fit` as the field
`par` in the outputted list}
}
\value{
a vector with randomly drawn observations
}
\description{
Generate random root diameter observations based on a multisegment power-law
fit result.
}
\examples{
# generate some data and fit
x1 <- rweibull(100, shape = 4, scale = 6)
par <- rootcount_fit(x1, 2)$par

# draw random values from fit
x2 <- rootcount_random(1000, par)

# plot cumulative data
c1 <- seq(0, 1, l = length(x1))
c2 <- seq(0, 1, l = length(x2))
plot(sort(x1), c1)
lines(sort(x2), c2, "l", col = "red")

}
