% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_range}
\alias{round_range}
\title{Determine rounded plot axis limits}
\usage{
round_range(x, lim = c(NA, NA), step = c(0.1, 0.2, 0.5, 1, 2, 5), ticks = 7)
}
\arguments{
\item{x}{values on axis that should be within the plot range}

\item{lim}{two-parameter vector with fixed axis limits. If \code{NA}, an axis
limit can be freely chosen}

\item{step}{array with possible step sizes. These values are relative to
the order of magnitude of the difference in input \code{x}, i.e. relative to
\code{10^floor(log10(max(x) - min(x)))}}

\item{ticks}{the maximum number of ticks that is acceptable}
}
\value{
a list with fields:
\itemize{
\item \code{lim}: two-parameter vector with min and max limit of the plot range
\item \code{breaks}: vector with plot breakpoints
}
}
\description{
Determine plot axis limits based on input values, preferred tick step sizes
and a maximum number of ticks
}
\examples{
# data
x <- 10*rnorm(100)

# basic range
z <- round_range(x)
plot(x, seq(length(x)), xlim = z$lim, xaxs = "i", xaxt = "n")
axis(side = 1, at = z$breaks)

# fix the lower bound
z <- round_range(x, lim = c(-100, NA))
plot(x, seq(length(x)), xlim = z$lim, xaxs = "i", xaxt = "n")
axis(side = 1, at = z$breaks)

}
