% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run a Shiny app that is part of this package}
\usage{
run_app(app_name = "powerlaw_customdata", package_name = "rootmech")
}
\arguments{
\item{app_name}{name of the application}

\item{package_name}{name of the current package}
}
\description{
This helper function launches a Shiny app that is part of this
R package
}
\examples{
\dontrun{run_app()}
}
