% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_fit}
\alias{weibull_fit}
\title{Weibull likelihood fitting}
\usage{
weibull_fit(
  x,
  weights = rep(1, length(x)),
  method = "bisection",
  range = c(-1, 1),
  start = NULL
)
}
\arguments{
\item{x}{array with observations}

\item{weights}{weighting factor for each observation in \code{x}}

\item{method}{choose \code{newton} for gradient descent solving, using the
\code{rootSolve::multiroot()} function, or \code{bisection} for bisection root
solving algorithm using the \code{stats::uniroot()} function.}

\item{range}{two-value array to add to best guess, to define initial
interval for bisection algorithm}

\item{start}{initial guess for weibull shape parameter. If not defined, an
estimate is made based on fitting of the linearised cumulative density
distribution}
}
\value{
a list containing the fields
\itemize{
\item \code{loglikelihood}: the log-likelihood score of the fit
\item \code{shape}: fitted weibull shape parameter
\item \code{scale}: fitted weibull scale parameter
}
}
\description{
Determine best fitting Weibull distribution using weighted loglikelihood
fitting.
}
\examples{
x <- stats::rweibull(50, 4, 2)
weibull_fit(x)

}
