% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_initialguess}
\alias{weibull_initialguess}
\title{Initial guess for weibull distribution fitting}
\usage{
weibull_initialguess(x, weights = rep(1, length(x)))
}
\arguments{
\item{x}{array with observations}

\item{weights}{weighting factor for each observation in \code{x}}
}
\value{
initial guess for weibull shape parameter
}
\description{
Return an initial guess for the Weibull shape parameter, to be used as a
starting point in function `weibull_fit()`.

The estimate is made using linear regression on the linearlised cumulative
density function, i.e by fitting

  log(-log(1 - P)) = shape*log(x) - shape*log(scale)

where P is the cumulative density for each value x
}
\keyword{internal}
