% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_loglikelihood}
\alias{weibull_loglikelihood}
\title{Loglikelihood of Weibull fit}
\usage{
weibull_loglikelihood(par, x, weights = rep(1, length(x)), deriv = 0)
}
\arguments{
\item{par}{vector with Weibull shape and scale parameters}

\item{x}{array with observations}

\item{weights}{weighting factor for each observation in \code{x}}
}
\value{
loglikelihood
}
\description{
Returns weighted loglikelihood function of Weibull fit or its
derivatives
}
\keyword{internal}
