% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_root}
\alias{weibull_root}
\title{Root to solve for Weibull fitting}
\usage{
weibull_root(shape, x, weights = rep(1, length(x)))
}
\arguments{
\item{shape}{shape parameter}

\item{x}{array with observations}

\item{weights}{weighting factor for each observation in \code{x}}
}
\value{
root
}
\description{
Root equation to solve for Weibull fitting
}
\keyword{internal}
