% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_root_jacobian}
\alias{weibull_root_jacobian}
\title{Jacobian of root to solve for Weibull fitting}
\usage{
weibull_root_jacobian(shape, x, weights = rep(1, length(x)))
}
\arguments{
\item{shape}{shape parameter}

\item{x}{array with observations}

\item{weights}{weighting factor for each observation in \code{x}}
}
\value{
derivative of function `weibull_root()` with respect to input
  argument `shape`
}
\description{
Jacobian of root equation to solve for Weibull fitting.
}
\examples{
# parameters
shape <- 4

# create some data
x <- stats::rweibull(50, shape, 2)
w <- stats::runif(length(x))

# check derivative - compare analytical and numerical solutions
eps <- 1e-6
v0 <- weibull_root(shape, x, weights = w)
v1 <- weibull_root(shape + eps, x, weights = w)
(v1 - v0)/eps
weibull_root_jacobian(shape, x, weights = w)

}
