% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_survival_nls_fit}
\alias{weibull_survival_nls_fit}
\title{Fit a weibull survival function using non-linear least-squares regreggion}
\usage{
weibull_survival_nls_fit(x)
}
\arguments{
\item{x}{vector with observations}
}
\value{
dataframe with fitting parameters `shape` and `scale`
}
\description{
Fit a Weibull survival function to a series of `x` observations. The
Weibull parameters are found by minimising the sum of squared residuals,
using R's `optim()` function.
}
\examples{
# generate data and fit
x <- sort(stats::rweibull(25, 4, 6))
ft1 <- weibull_survival_nls_fit(x)

# experimental data
n <- length(x)
y <- seq(1 - 1/(2*n), 1/(2*n), l = n)

# likelihood fit
ft2 <- weibull_fit(x)

# create prediction lines
xp <- seq(min(x), max(x), l = 101)
yp1 <- 1 - pweibull(xp, ft1$shape, ft1$scale)
yp2 <- 1 - pweibull(xp, ft2$shape, ft2$scale)

# plot
plot(x, y)
lines(xp, yp1, col = "red")
lines(xp, yp2, col = "blue")

}
