% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_survival_nls_initialguess}
\alias{weibull_survival_nls_initialguess}
\title{Initial guess for `weibull_survival_nls_fit()` function}
\usage{
weibull_survival_nls_initialguess(
  x,
  y = seq(1 - 1/(2 * length(x)), 1/(2 * length(x)), l = length(x))
)
}
\arguments{
\item{x}{vector with observations}

\item{y}{cumulative values for each observation in `x`}
}
\value{
vector with guess for shape and scale parameter
}
\description{
Generate initial guess for `weibull_survival_nls_fit()` function. The
guess is made by linear fitting on the linearlised x versus y data
}
