% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_survival_nls_rss}
\alias{weibull_survival_nls_rss}
\title{Residual sum-of-squares for `weibull_survival_nls_fit()` optimalisation}
\usage{
weibull_survival_nls_rss(
  par,
  x,
  y = seq(1 - 1/(2 * length(x)), 1/(2 * length(x)), l = length(x))
)
}
\arguments{
\item{par}{vector with fitting values (Weibull shape and scale parameters)}

\item{x}{vector with observations}

\item{y}{cumulative values for each observation in `x`}
}
\value{
residual sum of squares (scalar)
}
\description{
Calculate the residual sum of squares for Weibull survival function
fitting.
}
