% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_survival_nls_rss_jacobian}
\alias{weibull_survival_nls_rss_jacobian}
\title{Jacobian of `weibull_survival_nls_rss()`}
\usage{
weibull_survival_nls_rss_jacobian(
  par,
  x,
  y = seq(1 - 1/(2 * length(x)), 1/(2 * length(x)), l = length(x))
)
}
\arguments{
\item{par}{vector with fitting values (Weibull shape and scale parameters)}

\item{x}{vector with observations}

\item{y}{cumulative values for each observation in `x`}
}
\value{
vector with derivatives with respect to each element in `par`
}
\description{
Returns the derivative of the function `weibull_survival_nls_rss()` with
respect to fitting parameters `par`
}
\examples{
# test derivative
par <- c(3, 5)
x <- rweibull(100, par[1], par[2])
eps <- 1e-6
z0 <- weibull_survival_nls_rss(par, x)
z1a <- weibull_survival_nls_rss(par + c(eps, 0), x)
z1b <- weibull_survival_nls_rss(par + c(0, eps), x)
(c(z1a, z1b) - z0)/eps
weibull_survival_nls_rss_jacobian(par, x)

}
