% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{power_integrate}
\alias{power_integrate}
\title{Integrate a power-law curve}
\usage{
power_integrate(b, x1, x2)
}
\arguments{
\item{b}{vector with powers}

\item{x1, x2}{vectors with lower and upper integration limits}
}
\value{
vector with integrands
}
\description{
Integrate a power law curve in the form:

  y(x) = x^b

between x=x1 and x=x2.

Function is vectorised.
}
\examples{
power_integrate(2.1, 2, 5)
stats::integrate(function(x) x^2.1, 2, 5)

}
