% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_integrate.R
\name{power_integrate_jacobian}
\alias{power_integrate_jacobian}
\title{Derivative of function `power_integrate()`}
\usage{
power_integrate_jacobian(power, lower, upper, multiplier = 1)
}
\arguments{
\item{power}{vector with powers}

\item{lower, upper}{vectors with lower and upper integration limits}

\item{multiplier}{(optional) vector with multipliers}
}
\value{
a list containing the derivatives for each input parameter. Has
  fields `power`, `lower` and `upper` and `multiplier`.
}
\description{
Generates the derivative of the results of the function `power_integrate()`
with respect to its input parameters `power`, `lower`, `upper` and
`multiplier`.

Function is vectorised.
}
\examples{
power <- -0.4
lower <- 1.5
upper <- 3.6
multiplier <- 2.1

power_integrate_jacobian(power, lower, upper, multiplier = multiplier)

I0 <- power_integrate(power, lower, upper, multiplier = multiplier)
eps <- 1e-6
(power_integrate(power + eps, lower, upper, multiplier = multiplier) - I0)/eps
(power_integrate(power, lower + eps, upper, multiplier = multiplier) - I0)/eps
(power_integrate(power, lower, upper + eps, multiplier = multiplier) - I0)/eps
(power_integrate(power, lower, upper, multiplier = multiplier + eps) - I0)/eps

}
