% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlaw_normal.R
\name{powerlaw_normal_freebetadelta_root_jacobian}
\alias{powerlaw_normal_freebetadelta_root_jacobian}
\title{Jacobian of function `powerlaw_normal_freebetadelta_root()`}
\usage{
powerlaw_normal_freebetadelta_root_jacobian(
  par,
  x,
  y,
  weights = rep(1, length(x))
)
}
\arguments{
\item{par}{vector with mean and standard deviation power-law exponents}

\item{x}{measured x-values (e.g. root diameters). These values are assumed
to be already normalised by a reference value to ensure a dimensionless
parameter.}

\item{y}{measured y-values (e.g. root tensile strength)}

\item{weights}{vector with weighting for each observation (default: 1)}
}
\value{
derivative of `powerlaw_normal_freebetadelta_root()` with respect to
  elements in `par`
}
\description{
Returns the derivative of the function `powerlaw_normal_freebetadelta_root()`
with respect to input argument `par`
}
\keyword{internal}
