% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootcount.R
\name{rootcount_fit}
\alias{rootcount_fit}
\title{Fit a multisegment power-law probability distribution}
\usage{
rootcount_fit(
  x,
  ns,
  xmin = min(x, na.rm = TRUE),
  xmax = max(x, na.rm = TRUE),
  weights = rep(1, length(x)),
  par0 = NULL
)
}
\arguments{
\item{x}{vector with observations}

\item{ns}{number of segments to use}

\item{xmin, xmax}{minimum and maximum values of `x` to use in the fit.
By default, `xmin = min(x)` and `xmax = max(x)` since this will maximise
the probability. These values may however be overwritten if required.}

\item{weights}{vector with weights for each observation. The individual
probabilities are raised to the power `weights`. By default, all
observations are weighted equally (`weights = 1`)}

\item{par0}{vector with user-defined initial guess for th vector of fitting
parameters (`ns - 1` breakpoints, followed by `ns` power coefficients).
If not defined, a simple guess is made}
}
\value{
a list with the loglikelihood (field `loglikelihood`) and a
  dataframe (field `par`) with the fitting results for each segment.
  Each segment is defined using the lower and upper x-limit (`xmin`, `xmax`),
  the power-law coefficient (`power`) and the total probability in the
  segment (`total`).
}
\description{
Fit a C^0 continuous, multisegment power-law probability distribution to a
sset of observations `x`. Observations can be weighted if required.
}
\examples{
x <- rweibull(100, 4, 6)
rootcount_fit(x, 2)

}
