% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull_fit}
\alias{weibull_fit}
\title{Fit weibull distribution to a series of observations}
\usage{
weibull_fit(
  x,
  shape = NULL,
  scale = NULL,
  mu = NULL,
  weights = rep(1, length(x)),
  shape_range = c(1, 100)
)
}
\arguments{
\item{x}{vector with observations}

\item{shape}{if not `NULL`, a fixed value for the shape parameter}

\item{scale}{if not `NULL`, a fixed value for the scale parameter}

\item{weights}{a vector with weights for each observation}

\item{shape_range}{default `uniroot` interval. It is assumed that the actual
shape parameter is somewhere in this range, otherwise the function will
not work}
}
\value{
dataframe with fields for `loglikelihood`, `shape` and `scale`
}
\description{
Fit a Weibull probability distribution to a series of observations. The
best fit is determined by using a root solve on the  derivative of the
loglikelihood function, using the default R function `uniroot()`
}
\examples{
shape <- 4
scale <- 8
x <- rweibull(1000, shape, scale)

ft <- weibull_fit(x, shape = 3, mu = mean(x))

hist(x, breaks = 25, freq = FALSE)
xp <- seq(0, max(x), l = 101)
yp <- dweibull(xp, ft$shape, ft$scale)
lines(xp, yp)

}
